### 2021-05-29 #3321 @whedon
[PRE REVIEW]: ClusterValidityIndices.jl: Batch and Incremental Metrics for Unsupervised Learning
**Submitting author:** @AP6YC (<a href="http://orcid.org/0000-0003-2442-8901">Sasha Petrenko</a>)
**Repository:** <a href="https://github.com/AP6YC/ClusterValidityIndices.jl" target ="_blank">https://github.com/AP6YC/ClusterValidityIndices.jl</a>
**Version:** 0.1.5
