### 2021-05-29 #3321 @arfon
While you're waiting, please fix up your paper (the affiliations aren't formatted properly). See the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) for how to structure your paper metadata. You can, then use the following commands (one at a time, as the first line of a new comment) to regenerate the PDF and check the references.
