### 2021-06-01 ab7a2e2
Hotfix affiliations in paper.md
