### 2021-06-08 #3321 @AP6YC
@arfon You've got it! I totally understand the backlog. I just appreciate any attention that you are able to give and when you are able to give it, especially in these strange and challenging times.
