### 2021-06-08 #3321 @AP6YC
I have made some changes to the affiliations field and will attempt to rerun here:
