+++ src/ClusterValidityIndices.jl
@@ -4,7 +4,11 @@ using Statistics: mean
 
 abstract type AbstractCVI end
 
+# All of the CVI modules
 include("CVI/CVI.jl")
+
+# CVI utility functions
+include("utils.jl")
 
 # Export all public names
 export
@@ -29,6 +33,15 @@ export
     param_batch!,
     evaluate!,
     get_icvi!,
-    get_cvi!
+    get_cvi!,
+
+    # CVI utilities
+    sort_cvi_data,
+    relabel_cvi_data
+
+    # Not exported
+    # get_cvi_data
+    # get_bernoulli_subset
+    # showtypetree
 
 end
