### 2021-07-01 3144505
Increase test util coverage
