### 2021-07-06 3dacd5b
Define explicit abstract type inputs, bump
