+++ src/CVI/GD53.jl
@@ -34,18 +34,18 @@ using LinearAlgebra
 The stateful information of the Generalized Dunn's Index 53 (GD53) CVI.
 """
 mutable struct GD53 <: AbstractCVI
-    dim::Int64
-    n_samples::Int64
-    mu_data::Array{Float64, 1}  # dim
-    n::Array{Int64, 1}          # dim
-    v::Array{Float64, 2}        # dim x n_clusters
-    CP::Array{Float64, 1}       # dim
-    G::Array{Float64, 2}        # dim x n_clusters
-    D::Array{Float64, 2}        # n_clusters x n_clusters
-    inter::Float64
-    intra::Float64
-    n_clusters::Int64
-    criterion_value::Float64
+    dim::Integer
+    n_samples::Integer
+    mu_data::RealVector     # dim
+    n::IntegerVector        # dim
+    v::RealMatrix           # dim x n_clusters
+    CP::RealVector          # dim
+    G::RealMatrix           # dim x n_clusters
+    D::RealMatrix           # n_clusters x n_clusters
+    inter::RealFP
+    intra::RealFP
+    n_clusters::Integer
+    criterion_value::RealFP
 end # GD53 <: AbstractCVI
 
 """
@@ -57,12 +57,12 @@ function GD53()
     GD53(
         0,                              # dim
         0,                              # n_samples
-        Array{Float64, 1}(undef, 0),    # mu_data
-        Array{Int64, 1}(undef, 0),      # n
-        Array{Float64, 2}(undef, 0, 0), # v
-        Array{Float64, 1}(undef, 0),    # CP
-        Array{Float64, 2}(undef, 0, 0), # G
-        Array{Float64, 2}(undef, 0, 0), # D
+        Array{RealFP, 1}(undef, 0),     # mu_data
+        Array{Integer, 1}(undef, 0),    # n
+        Array{RealFP, 2}(undef, 0, 0),  # v
+        Array{RealFP, 1}(undef, 0),     # CP
+        Array{RealFP, 2}(undef, 0, 0),  # G
+        Array{RealFP, 2}(undef, 0, 0),  # D
         0.0,                            # inter
         0.0,                            # intra
         0,                              # n_clusters
@@ -70,21 +70,24 @@ function GD53()
     )
 end # GD53()
 
-function setup!(cvi::GD53, sample::Array{T, 1}) where {T<:Real}
+"""
+    setup!(cvi::GD53, sample::Array{T, 1}) where {T<:RealFP}
+"""
+function setup!(cvi::GD53, sample::Array{T, 1}) where {T<:RealFP}
     # Get the feature dimension
     cvi.dim = length(sample)
     # Initialize the augmenting 2-D arrays with the correct feature dimension
     # NOTE: R is emptied and calculated in evaluate!, so it is not defined here
     cvi.v = Array{T, 2}(undef, cvi.dim, 0)
     cvi.G = Array{T, 2}(undef, cvi.dim, 0)
-end
+end # setup!(cvi::GD53, sample::Array{T, 1}) where {T<:RealFP}
 
 """
-    param_inc!(cvi::GD53, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}
+    param_inc!(cvi::GD53, sample::RealVector, label::Integer)
 
 Compute the Generalized Dunn's Index 53 (GD53) CVI incrementally.
 """
-function param_inc!(cvi::GD53, sample::Array{T, 1}, label::I) where {T<:Real, I<:Int}
+function param_inc!(cvi::GD53, sample::RealVector, label::Integer)
[…]
