+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"man/contributing/#Contributing","page":"Contributing","title":"Contributing","text":"","category":"section"},{"location":"man/contributing/","page":"Contributing","title":"Contributing","text":"To formally contribute to the package, please follow the usual branch pull request procedure:","category":"page"},{"location":"man/contributing/","page":"Contributing","title":"Contributing","text":"Fork the project.\nCreate your feature branch (git checkout -b my-new-feature).\nCommit your changes (git commit -am 'Added some feature').\nPush to the branch (git push origin my-new-feature).\nCreate a new GitHub pull request.","category":"page"},{"location":"man/contributing/","page":"Contributing","title":"Contributing","text":"If you simply have suggestions for improvement, Sasha Petrenko (<sap625@mst.edu>) is the current developer and maintainer of the ClusterValidityIndices.jl package, so please feel free to reach out with thoughts and questions.","category":"page"},{"location":"man/full-index/#main-index","page":"Index","title":"Index","text":"","category":"section"},{"location":"man/full-index/","page":"Index","title":"Index","text":"Pages = [\"lib/public.md\"]","category":"page"},{"location":"man/full-index/","page":"Index","title":"Index","text":"Modules = [ClusterValidityIndices]","category":"page"},{"location":"man/full-index/#ClusterValidityIndices.CH","page":"Index","title":"ClusterValidityIndices.CH","text":"CH\n\nThe stateful information of the Calinski-Harabasz (CH) CVI.\n\n\n\n\n\n","category":"type"},{"location":"man/full-index/#ClusterValidityIndices.CH-Tuple{}","page":"Index","title":"ClusterValidityIndices.CH","text":"CH()\n\nDefault constructor for the Calinski-Harabasz (CH) Cluster Validity Index.\n\n\n\n\n\n","category":"method"},{"location":"man/full-index/#ClusterValidityIndices.DB","page":"Index","title":"ClusterValidityIndices.DB","text":"DB\n\nThe stateful information of the Davies-Bouldin (DB) CVI.\n\n\n\n\n\n","category":"type"},{"location":"man/full-index/#ClusterValidityIndices.DB-Tuple{}","page":"Index","title":"ClusterValidityIndices.DB","text":"DB()\n\nDefault constructor for the Davies-Bouldin (DB) Cluster Validity Index.\n\n\n\n\n\n","category":"method"},{"location":"man/full-index/#ClusterValidityIndices.GD43","page":"Index","title":"ClusterValidityIndices.GD43","text":"GD43\n\nThe stateful information of the Generalized Dunn's Index 43 (GD43) CVI.\n\n\n\n\n\n","category":"type"},{"location":"man/full-index/#ClusterValidityIndices.GD43-Tuple{}","page":"Index","title":"ClusterValidityIndices.GD43","text":"GD43()\n\nDefault constructor for the Generalized Dunn's Index 43 (GD43) Cluster Validity Index.\n\n\n\n\n\n","category":"method"},{"location":"man/full-index/#ClusterValidityIndices.GD53","page":"Index","title":"ClusterValidityIndices.GD53","text":"GD53\n\nThe stateful infor
[…]
