### 2021-07-08 4de9e6d
Remove extraneous methods in CVI.jl
