### 2021-07-08 bab8905
Move porcelain functions to common.jl
