+++ Project.toml
@@ -11,7 +11,7 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
-ProgressBars = "1.0"
+ProgressBars = "0.7, 0.8, 1"
 julia = "1"
 
 [extras]
