### 2021-07-22 #3321 @AP6YC
@adi3 As in the other thread, thank you for pinging me! I do sincerely apologize for the great delay.

Because this project requires similar domain knowledge, I am listing the same 6 reviewers as in AdaptiveResonance.jl (listed by username):

- ChrisRackauckas
- malmaud
- degleris1
- rMassimiliano
- aaronpeikert
- szkafander

If you wish, I am more than happy to select a separate group of reviewers. Thank you again for your patience.
