### 2021-07-23 #3527 @whedon
[REVIEW]: ClusterValidityIndices.jl: Batch and Incremental Metrics for Unsupervised Learning
**Submitting author:** @AP6YC (<a href="http://orcid.org/0000-0003-2442-8901">Sasha Petrenko</a>)
**Repository:** https://github.com/AP6YC/ClusterValidityIndices.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.6.3
