### 2021-08-27 cb8c105
Standardize setup! usage for type stability
