+++ src/CVI/WB.jl
@@ -71,16 +71,16 @@ function WB()
 end # WB()
 
 """
-    setup!(cvi::WB, sample::Array{T, 1}) where {T<:Real}
+    setup!(cvi::WB, sample::Vector{T}) where {T<:Real}
 """
-function setup!(cvi::WB, sample::Array{T, 1}) where {T<:Real}
+function setup!(cvi::WB, sample::Vector{T}) where {T<:Real}
     # Get the feature dimension
     cvi.dim = length(sample)
     # Initialize the augmenting 2-D arrays with the correct feature dimension
     # NOTE: R is emptied and calculated in evaluate!, so it is not defined here
     cvi.v = Array{T, 2}(undef, cvi.dim, 0)
     cvi.G = Array{T, 2}(undef, cvi.dim, 0)
-end # setup!(cvi::WB, sample::Array{T, 1}) where {T<:Real}
+end # setup!(cvi::WB, sample::Vector{T}) where {T<:Real}
 
 """
     param_inc!(cvi::WB, sample::RealVector, label::Integer)
