+++ src/CVI/GD43.jl
@@ -71,16 +71,16 @@ function GD43()
 end # GD43()
 
 """
-    setup!(cvi::GD43, sample::Array{T, 1}) where {T<:Real}
+    setup!(cvi::GD43, sample::Vector{T}) where {T<:RealFP}
 """
-function setup!(cvi::GD43, sample::Array{T, 1}) where {T<:RealFP}
+function setup!(cvi::GD43, sample::Vector{T}) where {T<:RealFP}
     # Get the feature dimension
     cvi.dim = length(sample)
     # Initialize the augmenting 2-D arrays with the correct feature dimension
     # NOTE: R is emptied and calculated in evaluate!, so it is not defined here
     cvi.v = Array{T, 2}(undef, cvi.dim, 0)
     cvi.G = Array{T, 2}(undef, cvi.dim, 0)
-end # setup!(cvi::GD43, sample::Array{T, 1}) where {T<:Real}
+end # setup!(cvi::GD43, sample::Vector{T}) where {T<:RealFP}
 
 """
     param_inc!(cvi::GD43, sample::RealVector, label::Integer)
