+++ src/CVI/XB.jl
@@ -65,15 +65,15 @@ function XB()
 end # XB()
 
 """
-    setup!(cvi::XB, sample::Array{T, 1}) where {T<:Real}
+    setup!(cvi::XB, sample::Vector{T}) where {T<:RealFP}
 """
-function setup!(cvi::XB, sample::Array{T, 1}) where {T<:Real}
+function setup!(cvi::XB, sample::Vector{T}) where {T<:RealFP}
     # Get the feature dimension
     cvi.dim = length(sample)
     # Initialize the 2-D arrays with the correct feature dimension
     cvi.v = Array{T, 2}(undef, cvi.dim, 0)
     cvi.G = Array{T, 2}(undef, cvi.dim, 0)
-end # setup!(cvi::XB, sample::Array{T, 1}) where {T<:Real}
+end # setup!(cvi::XB, sample::Vector{T}) where {T<:RealFP}
 
 """
     param_inc!(cvi::XB, sample::RealVector, label::Integer)
