+++ src/CVI/GD53.jl
@@ -71,16 +71,16 @@ function GD53()
 end # GD53()
 
 """
-    setup!(cvi::GD53, sample::Array{T, 1}) where {T<:RealFP}
+    setup!(cvi::GD53, sample::Vector{T}) where {T<:RealFP}
 """
-function setup!(cvi::GD53, sample::Array{T, 1}) where {T<:RealFP}
+function setup!(cvi::GD53, sample::Vector{T}) where {T<:RealFP}
     # Get the feature dimension
     cvi.dim = length(sample)
     # Initialize the augmenting 2-D arrays with the correct feature dimension
     # NOTE: R is emptied and calculated in evaluate!, so it is not defined here
     cvi.v = Array{T, 2}(undef, cvi.dim, 0)
     cvi.G = Array{T, 2}(undef, cvi.dim, 0)
-end # setup!(cvi::GD53, sample::Array{T, 1}) where {T<:RealFP}
+end # setup!(cvi::GD53, sample::Vector{T}) where {T<:RealFP}
 
 """
     param_inc!(cvi::GD53, sample::RealVector, label::Integer)
