+++ Project.toml
@@ -1,7 +1,7 @@
 name = "ClusterValidityIndices"
 uuid = "2fefd308-f647-4698-a2f0-e90cfcbca9ad"
 authors = ["Sasha Petrenko <sap625@mst.edu>"]
-version = "0.2.0"
+version = "0.2.1"
 
 [deps]
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"

[…]
