### 2021-09-06 #3527 @adi3
@malmaud @rMassimiliano - seems you are yet to accept the invite to be part of the JOSS org which will allow you to proceed through your checklist. Could you please let me know if I can help you sort out any trouble you're facing? Thanks!
