+++ src/common.jl
@@ -23,6 +23,13 @@ const IntegerMatrix{T<:Integer} = AbstractArray{T, 2}
 # Specifically floating-point aliases
 const RealFP = Union{Float32, Float64}
 
+# Internal label mapping for incremental CVIs
+const LabelMap = Dict{Int, Int}
+
+# -------------------------------------------
+# Methods
+# -------------------------------------------
+
 """
     get_icvi!(cvi::T, x::Array{N, 1}, y::M) where {T<:AbstractCVI, N<:Real, M<:Int}
 
@@ -54,3 +61,21 @@ function get_cvi!(cvi::T, x::Array{N, 2}, y::Array{M, 1}) where {T<:AbstractCVI,
     # Return that value
     return cvi.criterion_value
 end # get_cvi!(cvi::T, x::Array{N, 2}, y::Array{M, 1}) where {T<:AbstractCVI, N<:Real, M<:Int}
+
+"""
+    get_internal_label!(label_map::LabelMap, label::Int)
+
+Get the internal label and update the label map if the label is new.
+"""
+function get_internal_label!(label_map::LabelMap, label::Int)
+    # If the label map contains the key, return that internal label
+    if haskey(label_map, label)
+        internal_label = label
+    # Otherwise, increment the internal label to preserve monotonicity and store
+    else
+        internal_label = length(label_map) + 1
+        label_map[label] = internal_label
+    end
+
+    return internal_label
+end # get_internal_label!(label_map::LabelMap, label::Int)
