### 2021-09-29 3c35047
Implement label map for all CVIs, minor bump
