+++ src/CVI/DB.jl
@@ -28,6 +28,7 @@ using Statistics
 The stateful information of the Davies-Bouldin (DB) CVI.
 """
 mutable struct DB <: AbstractCVI
+    label_map::LabelMap
     dim::Integer
     n_samples::Integer
     mu_data::RealVector         # dim
@@ -49,6 +50,7 @@ Default constructor for the Davies-Bouldin (DB) Cluster Validity Index.
 """
 function DB()
     DB(
+        LabelMap(),                     # label_map
         0,                              # dim
         0,                              # n_samples
         Array{RealFP, 1}(undef, 0),     # mu_data
@@ -82,6 +84,9 @@ end # setup!(cvi::DB, sample::Vector{T}) where {T<:RealFP}
 Compute the Davies-Bouldin (DB) CVI incrementally.
 """
 function param_inc!(cvi::DB, sample::RealVector, label::Integer)
+    # Get the internal label
+    i_label = get_internal_label!(cvi.label_map, label)
+
     n_samples_new = cvi.n_samples + 1
     if isempty(cvi.mu_data)
         mu_data_new = sample
@@ -90,7 +95,7 @@ function param_inc!(cvi::DB, sample::RealVector, label::Integer)
         mu_data_new = (1 - 1/n_samples_new) .* cvi.mu_data + (1/n_samples_new) .* sample
     end
 
-    if label > cvi.n_clusters
+    if i_label > cvi.n_clusters
         n_new = 1
         v_new = sample
         CP_new = 0.0
@@ -105,8 +110,8 @@ function param_inc!(cvi::DB, sample::RealVector, label::Integer)
             for jx = 1:cvi.n_clusters
                 d_column_new[jx] = sum((v_new - cvi.v[:, jx]).^2)
             end
-            D_new[:, label] = d_column_new
-            D_new[label, :] = transpose(d_column_new)
+            D_new[:, i_label] = d_column_new
+            D_new[i_label, :] = transpose(d_column_new)
         end
         # Update 1-D parameters with a push
         cvi.n_clusters += 1
@@ -118,29 +123,29 @@ function param_inc!(cvi::DB, sample::RealVector, label::Integer)
         cvi.G = [cvi.G G_new]
         cvi.D = D_new
     else
-        n_new = cvi.n[label] + 1
-        v_new = (1 - 1/n_new) .* cvi.v[:, label] + (1/n_new) .* sample
-        delta_v = cvi.v[:, label] - v_new
+        n_new = cvi.n[i_label] + 1
+        v_new = (1 - 1/n_new) .* cvi.v[:, i_label] + (1/n_new) .* sample
+        delta_v = cvi.v[:, i_label] - v_new
         diff_x_v = sample .- v_new
-        CP_new = cvi.CP[label] + transpose(diff_x_v)*diff_x_v + cvi.n[label]*transpose(delta_v)*delta_v + 2*transpose(delta_v)*cvi.G[:, label]
-        G_new = cvi.G[:, label] + diff_x_v + cvi.n[label].*delta_v
+        CP_new = cvi.CP[i_label] + transpose(diff_x_v)*diff_x_v + cvi.n[i_label]*transpose(delta_v)*delta_v + 2*transpose(delta_v)*cvi.G[:, i_label]
+        G_new = cvi.G[:, i_label] + diff_x_v + cvi.n[i_label].*delta_v
         S_new = CP_new / n_new
         d_column_new = zeros(cvi.n_clusters)
         for jx = 1:cvi.n_clusters
-            # Skip the current label index
-            if jx == label
+            # Skip the current i_label index
+            if jx == i_label
                 continue
             end
[…]
