### 2021-11-02 #3527 @AP6YC
@adi3 Thanks for the ping; I'll tackle these shortly! I will update progress here as well as on the repo's issue pane.
