### 2021-11-02 AP6YC/ClusterValidityIndices.jl#3 @AP6YC
Quickstart docs
From https://github.com/openjournals/joss-reviews/issues/3527#issuecomment-946230234

Both the README and docs should have installation instructions for the package. They should also both contain a usage quickstart.
