### 2021-11-02 AP6YC/ClusterValidityIndices.jl#4 @AP6YC
CVI Docstrings
From https://github.com/openjournals/joss-reviews/issues/3527#issuecomment-946230234

All CVI/ICVIs should have much more detailed docstrings, perhaps linking to their relevant papers. At minimum, they should contain simple usage and document their arguments.
