### 2021-11-02 AP6YC/ClusterValidityIndices.jl#5 @AP6YC
State of the field
From https://github.com/openjournals/joss-reviews/issues/3527#issuecomment-947094105

The state-of-the-art of CVI implementations is missing in the paper. At minimum, several existing software implementations should be mentioned, which could be tied to the statement of need of the package.
