### 2021-11-02 AP6YC/ClusterValidityIndices.jl#6 @AP6YC
JOSS paper typos
From https://github.com/openjournals/joss-reviews/issues/3527#issuecomment-947094105

Typos in that comment should be addressed in the JOSS paper.
