### 2021-11-02 AP6YC/ClusterValidityIndices.jl#7 @AP6YC
Examples and introductions
From https://github.com/openjournals/joss-reviews/issues/3527#issuecomment-954321786

This comment illuminates many critical documentation needs:

- There are errors in the documentation that should be fixed from this comment.
- An conceptual introduction to CVIs is needed.
- The index needs to be cleaned up to cleanly demonstrate the API.
- The existing quickstart is confusing about nomenclature.
- Examples in the quickstart should be expanded for clarity.
- Performance of the algorithms should be mentioned (time complexity, memory usage, etc.).
