### 2021-11-02 AP6YC/ClusterValidityIndices.jl#8 @AP6YC
PkgEval failure
From https://github.com/openjournals/joss-reviews/issues/3527#issuecomment-954321786

Julia PkgEval is currently failing after previously passing, which should be resolved before the JOSS paper can be accepted.
