### 2021-11-03 29e076e
Add acknowledgements, authors, and license section in README
