### 2021-11-03 9cad0ca
Update porcelain methods to use type aliases
