### 2021-11-03 59ca029
Consolidate package dependencies into module def
