+++ src/CVI/XB.jl
@@ -20,8 +20,6 @@ cluster validity indices for performance monitoring of streaming data clustering
 Int. J. Intell. Syst., pp. 1–23, 2018.
 """
 
-using Statistics
-
 """
     XB
 
