### 2021-11-03 5bc357a
Add simple quickstart, shuffle sections
