### 2021-11-03 1313654
Add contributing to README and flesh out contributing doc
