### 2021-11-29 AP6YC/ClusterValidityIndices.jl#3 @AP6YC
PR #9 adds a quickstart and installation instructions to both the docs and the README.
