### 2021-11-30 AP6YC/ClusterValidityIndices.jl#8 @AP6YC
A previous update seems to have solved this bug, though I do not know which did.

![image](https://user-images.githubusercontent.com/17498866/143968836-8ad5cce0-abfa-4d80-b24a-db4fb56e3ef6.png)
