### 2021-11-30 #3527 @AP6YC
Sure thing! The timeline is ASAP, given that I have slipped a personal deadline for end of the month. After listing the PRs that fix the issues above, I will tackle the remaining items in the checklist.
