### 2021-11-30 #3527 @AP6YC
[PR 9](https://github.com/AP6YC/ClusterValidityIndices.jl/pull/9) solve [Issue 3](https://github.com/AP6YC/ClusterValidityIndices.jl/issues/3): quickstart and usage in the docs and README. This PR also significantly fleshes out the contributing page for the project after implementing GitFlow, also detailing how to contribute modules to the project.
