### 2021-11-30 022136c
Add relevant references to every CVI docstring
