+++ src/CVI/DB.jl
@@ -23,7 +23,12 @@ Int. J. Intell. Syst., pp. 1–23, 2018.
 """
     DB
 
-The stateful information of the Davies-Bouldin (DB) CVI.
+The stateful information of the Davies-Bouldin (DB) Cluster Validity Index.
+
+# References
+1. D. L. Davies and D. W. Bouldin, "A cluster separation measure," IEEE Transaction on Pattern Analysis and Machine Intelligence, vol. 1, no. 2, pp. 224–227, Feb. 1979.
+2. M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML]. [Online].
+3. M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1–23, 2018.
 """
 mutable struct DB <: CVI
     label_map::LabelMap
