### 2021-12-02 62ad490
Begin common docstring dev
