### 2021-12-02 b96e2b0
Strip individual cvi docstrings
