+++ src/CVI/WB.jl
@@ -89,11 +89,6 @@ function setup!(cvi::WB, sample::Vector{T}) where {T<:Real}
     cvi.G = Array{T, 2}(undef, cvi.dim, 0)
 end # setup!(cvi::WB, sample::Vector{T}) where {T<:Real}
 
-"""
-    param_inc!(cvi::WB, sample::RealVector, label::Integer)
-
-Compute the WB-Index (WB) CVI incrementally.
-"""
 function param_inc!(cvi::WB, sample::RealVector, label::Integer)
     # Get the internal label
     i_label = get_internal_label!(cvi.label_map, label)
@@ -136,11 +131,6 @@ function param_inc!(cvi::WB, sample::RealVector, label::Integer)
     cvi.SEP = [cvi.n[ix] * sum((cvi.v[:, ix] - cvi.mu).^2) for ix=1:cvi.n_clusters]
 end # param_inc!(cvi::WB, sample::RealVector, label::Integer)
 
-"""
-    param_batch!(cvi::WB, data::RealMatrix, labels::IntegerVector)
-
-Compute the WB-Index (WB) CVI in batch.
-"""
 function param_batch!(cvi::WB, data::RealMatrix, labels::IntegerVector)
     cvi.dim, cvi.n_samples = size(data)
     # Take the average across all samples, but cast to 1-D vector
@@ -162,11 +152,6 @@ function param_batch!(cvi::WB, data::RealMatrix, labels::IntegerVector)
     end
 end # param_batch!(cvi::WB, data::RealMatrix, labels::IntegerVector)
 
-"""
-    evaluate!(cvi::WB)
-
-Compute the criterion value of the WB-Index (WB) CVI.
-"""
 function evaluate!(cvi::WB)
     # Within group sum of scatters
     cvi.WGSS = sum(cvi.CP)

[…]
