### 2021-12-03 5ab1309
Add datadeps env variable to docs build
