+++ .github/workflows/Documentation.yml
@@ -28,5 +28,7 @@ jobs:
       - name: Build and deploy
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # For authentication with GitHub Actions token
+          # Set this to download test data outside of repo
+          DATADEPS_ALWAYS_ACCEPT: true
           # DOCUMENTER_KEY: ${{ secrets.DOCUMENTER_KEY }} # For authentication with SSH deploy key
         run: julia --project=docs/ docs/make.jl
