### 2021-12-06 a73bddc
Add more docstring elements to common methods
