+++ src/common.jl
@@ -105,6 +105,22 @@ end # get_internal_label!(label_map::LabelMap, label::Int)
     param_inc!(cvi::CVI, sample::RealVector, label::Integer)
 
 Compute the CVI parameters incrementally.
+
+This method updates only internal parameters of the ICVI algorithm incrementally.
+When the criterion value itself is needed, use `evaluate!` and extract it from `cvi.criterion_value`.
+
+# Arguments
+- `cvi::CVI`: the stateful information of the CVI/ICVI algorithm.
+- `sample::RealVector`: a vector of features used in the external clustering algorithm.
+- `label::Integer`: the label that the external clustering algorithm prescribed to the `sample`.
+
+# Examples
+```julia-repl
+julia> my_cvi = CH()
+julia> data = load_some_data()
+julia> labels = my_cluster_algorithm(data)
+julia> param_inc!(my_cvi, data[:, 1], labels[1])
+```
 """
 param_inc!(cvi::CVI, sample::RealVector, label::Integer)
 
@@ -112,6 +128,22 @@ param_inc!(cvi::CVI, sample::RealVector, label::Integer)
     param_batch!(cvi::CVI, data::RealMatrix, labels::IntegerVector)
 
 Compute the CVI parameters in batch.
+
+This method updates only the internal parameters of the CVI algorithm in batch.
+When the criterion value itself is needed, use `evaluate!` and extract it from `cvi.criterion_value`.
+
+# Arguments
+- `cvi::CVI`: the stateful information of the CVI/ICVI algorithm.
+- `data::RealMatrix`: a matrix of data where rows are features and columns are samples, used in the external clustering algorithm.
+- `labels::IntegerVector`: a vector of labels that the external clustering algorithm prescribed to each column in `data`.
+
+# Examples
+```julia-repl
+julia> my_cvi = CH()
+julia> data = load_some_data()
+julia> labels = my_cluster_algorithm(data)
+julia> param_batch!(my_cvi, data, labels)
+```
 """
 param_batch!(cvi::CVI, data::RealMatrix, labels::IntegerVector)
 
@@ -119,5 +151,21 @@ param_batch!(cvi::CVI, data::RealMatrix, labels::IntegerVector)
     evaluate!(cvi::CVI)
 
 Compute the criterion value of the CVI.
+
+After computation, the resulting criterion value can be extracted from `cvi.criterion_value`.
+The criterion value is a function of the CVI/ICVI internal parameters, so at least two classes (i.e., unique labels) must be presented to the CVI in `param_inc!` or `param_batch!` before a non-zero value is returned.
+
+# Arguments
+- `cvi::CVI`: the stateful information of the CVI/ICVI to use for computing the criterion value.
+
+# Examples
+```julia-repl
+julia> my_cvi = CH()
+julia> data = load_some_data()
+julia> labels = my_cluster_algorithm(data)
+julia> param_batch!(my_cvi, data, labels)
+julia> evaluate!(my_cvi)
+julia> my_criterion_value = my_cvi.criterion_value
+```
 """
 evaluate!(cvi::CVI)
