+++ src/common.jl
@@ -48,36 +48,76 @@ const LabelMap = Dict{Int, Int}
 # -------------------------------------------
 
 """
-    get_icvi!(cvi::CVI, x::RealVector, y::Integer)
+    get_icvi!(cvi::CVI, sample::RealVector, label::Integer)
 
-Porcelain: update and compute the criterion value incrementally and return it.
+Compute and return the criterion value incrementally.
+
+# Arguments
+- `cvi::CVI`: the stateful information of the ICVI providing the criterion value.
+- `sample::RealVector`: the sample provided to the external. clustering algorithm.
+- `label::Integer`: the label prescribed to the `sample` by the external clustering algorithm.
+
+# Examples
+```julia
+# Create a new CVI object
+my_cvi = CH()
+# Load in data from some external source
+data = load_some_data()
+# Cluster the data into a set of labels as an integer vector
+labels = my_cluster_algorithm(data)
+# Iteratively compute and extract the criterion value at every step
+n_samples = length(labels)
+criterion_values = zeros(n_samples)
+for ix = 1:n_samples
+    sample = data[:, ix]
+    label = labels[ix]
+    criterion_values[ix] = get_icvi!(my_cvi, sample, label)
+end
+```
 """
-function get_icvi!(cvi::CVI, x::RealVector, y::Integer)
+function get_icvi!(cvi::CVI, sample::RealVector, label::Integer)
     # Update the ICVI parameters
-    param_inc!(cvi, x, y)
+    param_inc!(cvi, sample, label)
 
     # Compute the criterion value
     evaluate!(cvi)
 
     # Return that value
     return cvi.criterion_value
-end # get_icvi!(cvi::CVI, x::RealVector, y::Integer)
+end # get_icvi!(cvi::CVI, sample::RealVector, label::Integer)
 
 """
-    get_cvi!(cvi::CVI, x::RealMatrix, y::IntegerVector)
+    get_cvi!(cvi::CVI, data::RealMatrix, labels::IntegerVector)
 
-Porcelain: update compute the criterion value in batch and return it.
+Compute and return the criterion value in batch mode.
+
+# Arguments
+- `cvi::CVI`: the stateful information of the CVI providing the criterion value.
+- `data::RealMatrix`: a matrix of data, columns as samples and rows as features, used in the external clustering process.
+- `labels::IntegerVector`: a vector of integers representing labels prescribed to the `data` by the external clustering algorithm.
+
+# Examples
+```julia
+# Create a new CVI object
+my_cvi = CH()
+# Load in data from some external source
+data = load_some_data()
+# Cluster the data into a set of labels as an integer vector
+labels = my_cluster_algorithm(data)
+# Compute the final criterion value in batch mode
+criterion_value = get_cvi!(cvi, data, labels)
+```
 """
-function get_cvi!(cvi::CVI, x::RealMatrix, y::IntegerVector)
+function get_cvi!(cvi::CVI, data::RealMatrix, labels::IntegerVector)
     # Update the CVI parameters in batch
-    param_batch!(cvi, x, y)
+    param_batch!(cvi, data, labels)
 
     # Compute the criterion value
     evaluate!(cvi)
 
     # Return that value
     return cvi.criterion_value
-end # get_cvi!(cvi::CVI, x::RealMatrix, y::IntegerVector)
+end # get_cvi!(cvi::CVI, data::RealMatrix, labels::IntegerVector)
 
 """
     get_internal_label!(label_map::LabelMap, label::Int)
