### 2021-12-06 #3527 @AP6YC
@malmaud Partial progress has been made, though I wouldn't consider the entire project ready for another round of reviewing in the interest of respecting your time. I tried atomizing your comments into issues on the project's page to tackle one-by-one, and the ones that I close with a successful PR are ready for your review to make sure that they are to your satisfaction.
