### 2021-12-06 a767ac2
Rewrite full index, add logos
