+++ docs/make.jl
@@ -1,10 +1,24 @@
 using Documenter
 using ClusterValidityIndices
+using DemoCards
 
+# Generate the demo files
+# this is the relative path to docs/
+demopage, postprocess_cb, demo_assets = makedemos("examples")
+
+assets = [
+    joinpath("assets", "favicon.ico")
+]
+
+# if there are generated css assets, pass it to Documenter.HTML
+isnothing(demo_assets) || (push!(assets, demo_assets))
+
+# Make the documentation
 makedocs(
     modules=[ClusterValidityIndices],
     format=Documenter.HTML(
         prettyurls = get(ENV, "CI", nothing) == "true",
+        assets = assets,
     ),
     # format=Documenter.HTML(),
     pages=[
@@ -21,6 +35,9 @@ makedocs(
     authors="Sasha Petrenko",
     # assets=String[],
 )
+
+# 3. postprocess after makedocs
+postprocess_cb()
 
 deploydocs(
     repo="github.com/AP6YC/ClusterValidityIndices.jl.git",
