+++ docs/src/man/full-index.md
@@ -1,9 +1,39 @@
 # [Index](@id main-index)
 
+This page lists the core methods and types of the `ClusterValidityIndices.jl` package.
+The [Methods](@ref index-methods) section lists the public methods for the package that use the CVIs/ICVIs in [Types](@ref index-types).
+Each of these entries link to the docstrings in [Docs](@ref index-docs)
+
+## [Methods](@id index-methods)
+
 ```@index
-Pages = ["lib/public.md"]
+Modules = [ClusterValidityIndices]
+Order = [:function]
+Public = true
 ```
 
-```@autodocs
+## [Types](@id index-types)
+
+```@index
 Modules = [ClusterValidityIndices]
+Order = [:type]
+Public = true
+```
+
+## [Docs](@id index-docs)
+
+```@docs
+get_cvi!
+get_icvi!
+param_inc!
+param_batch!
+evaluate!
+CH
+cSIL
+DB
+GD43
+PS
+rCIP
+WB
+XB
 ```
