### 2021-12-07 99a5c33
Fix demo docs compilation
