+++ docs/examples/index.md
@@ -1,0 +1,9 @@
+# [Examples](@id examples)
+
+This section contains some examples using the `ClusterValidityIndices.jl` package with topics ranging from how to the internals of package work to practical examples on different datasets.
+
+---
+
+{{{democards}}}
+
+---
