+++ docs/Project.toml
@@ -1,5 +1,6 @@
 [deps]
 ClusterValidityIndices = "2fefd308-f647-4698-a2f0-e90cfcbca9ad"
+Clustering = "aaaa29a8-35af-508c-8bc3-b662a17a0fe5"
 DemoCards = "311a05b2-6137-4a5a-b473-18580a3d38b5"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 LiveServer = "16fef848-5104-11e9-1b77-fb7a48bbb589"
