+++ docs/examples/cvis/cvi-icvi.jl
@@ -1,0 +1,11 @@
+# ---
+# title: CVI-ICVI Comparison
+# id: cvi_icvi
+# cover: ../assets/logo.png
+# date: 2021-12-6
+# author: "[Sasha Petrenko](https://github.com/AP6YC)"
+# julia: 1.6
+# description: This demo illustrates how to use incremental training methods vs. batch training for all ART modules. This demo also demonstrates how the final results of each CVI and ICVI are equivalent.
+# ---
+
+# testing
