### 2021-12-07 a7e11ae
Implement more cvi example details
