### 2021-12-07 6abd11b
Document batch example
