+++ docs/Project.toml
@@ -1,4 +1,5 @@
 [deps]
+AdaptiveResonance = "3d72adc0-63d3-4141-bf9b-84450dd0395b"
 ClusterValidityIndices = "2fefd308-f647-4698-a2f0-e90cfcbca9ad"
 Clustering = "aaaa29a8-35af-508c-8bc3-b662a17a0fe5"
 DemoCards = "311a05b2-6137-4a5a-b473-18580a3d38b5"
