+++ docs/examples/cvis/cvi-example.jl
@@ -1,5 +1,5 @@
 # ---
-# title: CVI Example
+# title: CVI Simple Example
 # id: cvi_example
 # cover: ../assets/logo.png
 # date: 2021-12-6
