+++ docs/examples/cvis/icvi-comparison.jl
@@ -1,0 +1,11 @@
+# ---
+# title: ICVI Comparisons
+# id: icvi_comparison
+# cover: ../assets/logo.png
+# date: 2021-12-7
+# author: "[Sasha Petrenko](https://github.com/AP6YC)"
+# julia: 1.6
+# description: This demo illustrates the differing behavior of each CVI.
+# ---
+
+# testing
