### 2021-12-08 #3527 @AP6YC
[PR 10](https://github.com/AP6YC/ClusterValidityIndices.jl/pull/10) addresses [Issue 4](https://github.com/AP6YC/ClusterValidityIndices.jl/issues/4) and some more direct concerns, some of which may overlap:

Concerns of @malmaud
1. Correct package installation procedures.
2. Cleaned-up documentation page in the form of an index of user-level methods and types with fleshed-out docstrings.
3. Examples using DemoCards.jl in the Examples page that are generated with the docs and available as notebooks.

Concerns of @rMassimiliano
1. References to relevant papers in each CVI docstring.
2. Clarification of the CVI/ICVI nomenclature in the README quickstart and docs.
3. Clarification of the basic methods in the quickstart with types (more than just (...)).

Coming up next will be changes to the paper itself (typos, state-of-the-art, etc.), a background page on what CVIs are, and a solution to the stable docs link.
