### 2021-12-13 #3527 @adi3
@AP6YC - Thank you for your work so far! Can you please give us an estimated completion date for these issues so that the reviewers may continue with their checklist?
