### 2022-04-22 #3527 @AP6YC
All,

I sincerely apologize for my absence in this project and this thread, which despite life events I cannot excuse. I would like to thank @adi3 for reaching out to me via email to correct my neglect of this thread. To quickly conclude this review, and as a show of good faith to everyone's time, I have gone through this thread and listed out the remaining reviewers' concerns as a list of issues on the repo page:

- [ ] https://github.com/AP6YC/ClusterValidityIndices.jl/issues/7
- [ ] https://github.com/AP6YC/ClusterValidityIndices.jl/issues/6
- [ ] https://github.com/AP6YC/ClusterValidityIndices.jl/issues/5

This list will also accommodate any new concerns as I close things up. In summary, I will make immediate changes to the paper itself, such as comparisons to existing implementations for the state of the field concern, followed by cleaning the documentation (detailed in #6)

My next message will be to show that these last concerns are completed for the reviewers.
