+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @article{brito_da_silva_incremental_2020,
-	title = {Incremental {Cluster} {Validity} {Indices} for {Online} {Learning} of {Hard} {Partitions}: {Extensions} and {Comparative} {Study}},
+	title = {Incremental Cluster Validity Indices for Online Learning of Hard Partitions: Extensions and Comparative Study},
 	volume = {8},
 	doi = {10.1109/ACCESS.2020.2969849},
 	abstract = {Validation is one of the most important aspects of clustering, particularly when the user is designing a trustworthy or explainable system. However, most clustering validation approaches require batch calculation. This is an important gap because of the value of clustering in real-time data streaming and other online learning applications. Therefore, interest has grown in providing online alternatives for validation. This paper extends the incremental cluster validity index (iCVI) family by presenting incremental versions of Calinski-Harabasz (iCH), Pakhira-Bandyopadhyay-Maulik (iPBM), WB index (iWB), Silhouette (iSIL), Negentropy Increment (iNI), Representative Cross Information Potential (irCIP), Representative Cross Entropy (irH), and Conn\_Index (iConn\_Index). This paper also provides a thorough comparative study of correct, under- and over-partitioning on the behavior of these iCVIs, the Partition Separation (PS) index as well as four recently introduced iCVIs: incremental Xie-Beni (iXB), incremental Davies-Bouldin (iDB), and incremental generalized Dunn's indices 43 and 53 (iGD43 and iGD53). Experiments were carried out using a framework that was designed to be as agnostic as possible to the clustering algorithms. The results on synthetic benchmark data sets showed that while evidence of most under-partitioning cases could be inferred from the behaviors of the majority of these iCVIs, over-partitioning was found to be a more challenging problem, detected by fewer of them. Interestingly, over-partitioning, rather then under-partitioning, was more prominently detected on the real-world data experiments within this study. The expansion of iCVIs provides significant novel opportunities for assessing and interpreting the results of unsupervised lifelong learning in real-time, wherein samples cannot be reprocessed due to memory and/or application constraints.},
@@ -13,7 +13,7 @@
 }
 
 @article{Grossberg2013,
-    title = {{Adaptive Resonance Theory: How a brain learns to consciously attend, learn, and recognize a changing world}},
+    title = {Adaptive Resonance Theory: How a brain learns to consciously attend, learn, and recognize a changing world},
     year = {2013},
     journal = {Neural Networks},
     author = {Grossberg, Stephen},
@@ -28,7 +28,7 @@
 }
 
 @article{Grossberg1980,
-    title = {{How Does a Brain Build a Cognitive Code ?}},
+    title = {How Does a Brain Build a Cognitive Code?},
     year = {1980},
     journal = {Psychological Review},
     author = {Grossberg, Stephen},
@@ -42,7 +42,7 @@
 }
 
 @article{DaSilva2019,
-    title = {{A Survey of Adaptive Resonance Theory Neural Network Models for Engineering Applications}},
+    title = {A Survey of Adaptive Resonance Theory Neural Network Models for Engineering Applications},
     year = {2019},
     journal = {Neural Networks},
     author = {Brito da Silva, Leonardo Enzo and Elnabarawy, Islam and Wunsch, Donald C.},
