### 2022-04-28 AP6YC/ClusterValidityIndices.jl#6 @AP6YC
Manually closing after PR #12
