### 2022-04-28 AP6YC/ClusterValidityIndices.jl#14 @AP6YC
Stable docs
The stable documentation is not building in the `gh-pages` branch for each release (mentioned in https://github.com/openjournals/joss-reviews/issues/3527#issuecomment-954321786).
