### 2022-04-29 33a66b1
Adds DocStringExtensions dep and compat
