+++ Project.toml
@@ -4,6 +4,7 @@ authors = ["Sasha Petrenko <sap625@mst.edu>"]
 version = "0.3.2"
 
 [deps]
+DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 ProgressBars = "49802e3a-d2f1-5c88-81d8-b72133a6f568"
@@ -11,6 +12,7 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
+DocStringExtensions = "0.8"
 ProgressBars = "0.7, 0.8, 1"
 julia = "1"
 
