### 2022-05-03 e5138f4
Manually build and add v0.3.2 and v0.3.1 to docs
