+++ v0.3.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"examples/basic-examples/icvi-example.html#icvi_example","page":"ICVI Simple Example","title":"ICVI Simple Example","text":"","category":"section"},{"location":"examples/basic-examples/icvi-example.html","page":"ICVI Simple Example","title":"ICVI Simple Example","text":"(Image: Source code) (Image: notebook) (Image: compat) (Image: Author) (Image: Update time)","category":"page"},{"location":"examples/basic-examples/icvi-example.html#Overview","page":"ICVI Simple Example","title":"Overview","text":"","category":"section"},{"location":"examples/basic-examples/icvi-example.html","page":"ICVI Simple Example","title":"ICVI Simple Example","text":"This demo is a simple example of how to use CVIs incrementally. Here, we load a simple dataset and run a basic clustering algorithm to prescribe a set of clusters to the features. We will take advantage of the fact that we can compute a criterion value at every step by running the ICVI alongside an online clustering algorithm. This simple example demonstrates the usage of a single ICVI, but it may be substituted for any other ICVI in the ClusterValidityIndices.jl package.","category":"page"},{"location":"examples/basic-examples/icvi-example.html#Online-Clustering","page":"ICVI Simple Example","title":"Online Clustering","text":"","category":"section"},{"location":"examples/basic-examples/icvi-example.html#Data-Setup","page":"ICVI Simple Example","title":"Data Setup","text":"","category":"section"},{"location":"examples/basic-examples/icvi-example.html","page":"ICVI Simple Example","title":"ICVI Simple Example","text":"First, we must load all of our dependencies. We will load the ClusterValidityIndices.jl along with some data utilities and the Julia Clustering.jl package to cluster that data.","category":"page"},{"location":"examples/basic-examples/icvi-example.html","page":"ICVI Simple Example","title":"ICVI Simple Example","text":"using ClusterValidityIndices    # CVI/ICVI\nusing AdaptiveResonance         # DDVFA\nusing MLDatasets                # Iris dataset\nusing MLDataUtils               # Shuffling and splitting\nusing Printf                    # Formatted number printing\nusing Plots","category":"page"},{"location":"examples/basic-examples/icvi-example.html","page":"ICVI Simple Example","title":"ICVI Simple Example","text":"We will download the Iris dataset for its small size and benchmark use for clustering algorithms.","category":"page"},{"location":"examples/basic-examples/icvi-example.html","page":"ICVI Simple Example","title":"ICVI Simple Example","text":"Iris.download(i_accept_the_terms_of_use=true)\nfeatures, labels = Iris.features(), Iris.labels()","category":"page"},{"location":"examples/basic-examples/icvi-example.html","page":"ICVI Simple Example","title":"ICVI Simple Example","text":"Because the MLDatasets package gives us Iris labels as strings, we will use the MLDataUtils.convertlabel method with the MLLabelUtils.LabelEnc.Indices type to get a list of integers representing each class:","category":"page"},{"location":"examples/basic-examples/icvi-example.html","page":"ICVI Simple Example","title":"ICVI Simple Example","text":"labels = convertlabel(LabelEnc.Indices{Int}, labels)\nunique(labels)","category":"page"},{"location":"examples/basic-examples/icvi-example.html#ART-Online-Clustering","page":"ICVI Simple Example","title":"ART Online Clustering","text":"","category":"section"},{"location":"examples/basic-examples/icvi-example.html","page":"ICVI Simple Example","title":"ICVI Simple Example","text":"Adaptive Resonance Theory (ART) is a neurocognitive theory that is the basis of a class of
[…]
