+++ versions.js
@@ -1,6 +1,8 @@
 var DOC_VERSIONS = [
   "stable",
   "v0.3",
+  "v0.2",
+  "v0.1",
   "dev",
 ];
 var DOCUMENTER_NEWEST = "v0.3.2";
