+++ v0.1.5/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"man/examples.html#Examples","page":"Examples","title":"Examples","text":"","category":"section"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"There are examples for every structure in the package within the package's examples/ folder. The code for several of these examples is provided here.","category":"page"},{"location":"man/examples.html#CVI","page":"Examples","title":"CVI","text":"","category":"section"},{"location":"man/examples.html#Xie-Beni-(XB)","page":"Examples","title":"Xie-Beni (XB)","text":"","category":"section"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"cvi = XB()","category":"page"},{"location":"man/contributing.html#Contributing","page":"Contributing","title":"Contributing","text":"","category":"section"},{"location":"man/contributing.html","page":"Contributing","title":"Contributing","text":"To formally contribute to the package, please follow the usual branch pull request procedure:","category":"page"},{"location":"man/contributing.html","page":"Contributing","title":"Contributing","text":"Fork the project.\nCreate your feature branch (git checkout -b my-new-feature).\nCommit your changes (git commit -am 'Added some feature').\nPush to the branch (git push origin my-new-feature).\nCreate a new GitHub pull request.","category":"page"},{"location":"man/contributing.html","page":"Contributing","title":"Contributing","text":"If you simply have suggestions for improvement, Sasha Petrenko (<sap625@mst.edu>) is the current developer and maintainer of the ClusterValidityIndices.jl package, so please feel free to reach out with thoughts and questions.","category":"page"},{"location":"man/guide.html#Package-Guide","page":"Guide","title":"Package Guide","text":"","category":"section"},{"location":"man/guide.html#Installation","page":"Guide","title":"Installation","text":"","category":"section"},{"location":"man/guide.html","page":"Guide","title":"Guide","text":"The CVI package can be installed using the Julia package manager. From the Julia REPL, type ']' to enter the Pkg REPL mode and run","category":"page"},{"location":"man/guide.html","page":"Guide","title":"Guide","text":"pkg> add CVI","category":"page"},{"location":"man/guide.html","page":"Guide","title":"Guide","text":"Alternatively, it can be added to your environment in a script with","category":"page"},{"location":"man/guide.html","page":"Guide","title":"Guide","text":"using Pkg\r\nPkg.add(\"CVI\")","category":"page"},{"location":"man/guide.html#Quickstart","page":"Guide","title":"Quickstart","text":"","category":"section"},{"location":"man/guide.html","page":"Guide","title":"Guide","text":"This section provides a quick overview of how to use the project. For more detailed code usage, please see Usage.","category":"page"},{"location":"man/guide.html","page":"Guide","title":"Guide","text":"This project has several example scripts to demonstrate the functionality of CVIs in the ClusterValidityIndices.jl package. In ICVI-Examples/src/examples/, the scripts db.jl, ps.jl, and xb.jl demonstrate usage of the Davies-Boudin (DB), Partition Separation (PS), and Xie-Beni (XB) metrics, respectively.","category":"page"},{"location":"man/guide.html","page":"Guide","title":"Guide","text":"note: Note\nEach of these scripts must be run at the top level of the project to correctly point to the datasets. For example, they can be run in the shell withjulia src/examples/db.jlor in a Julia REPL session withinclude(\"src/examples/db.jl\")","category":"page"},{"location":"man/guide.html","page":"Guide","title":"Guide","text":"Three preprocessed datasets are provided under data/ to demonstrate the correct partiti
[…]
