+++ v0.1.6/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"man/examples.html#Examples","page":"Examples","title":"Examples","text":"","category":"section"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"There are examples for every structure in the package within the package's examples/ folder. The code for several of these examples is provided here.","category":"page"},{"location":"man/examples.html#CVI-Full-Usage-Example","page":"Examples","title":"CVI Full Usage Example","text":"","category":"section"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"Consider that you already have a dataset that is labeled by some clustering algorithm. This is not strictly necessary in practice, as the incremental variants of each CVI are designed to be able to run online alongside a clustering process, but we do so here for simplicity. We treat the labels here as clustering-prescribed labels rather than true supervised labels, and we treat the data as the samples that were used to cluster to those labels.","category":"page"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"We begin by loading the module and loading the data wherever it may be:","category":"page"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"# Load the CVI/ICVI module\r\nusing ClusterValidityIndices\r\n\r\n# Point to the data file\r\ndata_path = \"data/correct_partition.csv\"\r\n\r\n# Load the data and labels\r\ndata, labels = get_cvi_data(data_path)","category":"page"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"Because Julia is column-major in memory and our data samples are potentially large, we follow the Julia notation and treat the dimensions of data as [dim, n_samples].","category":"page"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"note: Note\nAs of ClusterValidityIndices.jl v0.1.5, all the CVIs assume that the labels are presented sequentially initially, starting with index 1 (e.g., 1, 1, 2, 2, 3, 2, 2, 1, 3, 4, 4 ...). You may repeat previously seen label indices, but skipping label indices (e.g., 1, 2, 4) results in undefined behavior. If your data does not accomodate this,we may circumvent this by relabelling the data monotonically withlabels = relabel_cvi_data(labels)","category":"page"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"We can get the number of samples from the length of the labels vector because each data sample corresponds to a label:","category":"page"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"# Get the number of samples for incremental iteration\r\nn_samples = length(labels)","category":"page"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"We are now ready to instantiate our CVI with default parameters. Because we have incremental and batch variants, we will instantiate two CVI models, train one sequentially and one in batch, and show that their results are equivalent. We will use the Davies-Bouldin CVI/ICVI as an example here, but all CVIs in this package have the same usage.","category":"page"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"# Create two containers for the batch and incremental CVIs\r\ncvi_i = DB()\r\ncvi_b = DB()","category":"page"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"We will preallocate an array for the criterion values of the incremental variant so that we can populate it iteratively.","category":"page"},{"location":"man/examples.html","page":"Examples","title":"Examples","text":"# Create some storage for our criterion values\r\ncriterion_values_i = zeros(n_samples)","catego
[…]
