### 2022-05-03 1a62d43
Update CI for latest and LTS Julia, macOS, and x86
