### 2022-05-03 ab563db
Update get_bernoulli_subset with module types
