### 2022-05-03 ee33d59
Update rCIP to use module type Float
