### 2022-05-03 49ae016
Update get_cvi_data to use Int conversion
