+++ test/utils.jl
@@ -26,7 +26,7 @@ function get_cvi_data(data_file::String)
     data = readdlm(data_file, ',')
     data = permutedims(data)
     train_x = data[1:2, :]
-    train_y = convert(Array{Int64}, data[3, :])
+    train_y = convert(Array{Int}, data[3, :])
 
     return train_x, train_y
 end # get_cvi_data(data_file::String)
