### 2022-05-03 550293a
Try accepting real in bernoulli subset
