### 2022-05-04 b906957
Update appveyor script
