### 2022-05-04 8890755
Update README and contributing guide
