### 2022-05-04 61a55d5
Correct typos in basic example file
