### 2022-05-04 eb583d7
Remove old example page, fix refs
