+++ docs/src/index.md
@@ -31,6 +31,6 @@ Depth = 1
 
 The [Background](@ref) provides an overview of the problem statement of CVIs and what they are theoretically, while [Basic Example](@ref) steps through an single example workflow.
 
-The [Package Guide](@ref) provides a tutorial to the full usage of the package, while [Examples](@ref) gives many sample workflows using a variety of CVI modules.
+The [Package Guide](@ref) provides a tutorial to the full usage of the package, while [Examples](@ref examples) gives many sample workflows using a variety of CVI modules.
 
 Instructions on how to contribute to the package are found in [Contributing](@ref), and docstrings for every element of the package is listed in the [Index](@ref main-index).
