### 2022-05-04 7cea559
Add index entries, docstring CVI abstract type
