### 2022-05-04 2e9af33
Comment out unreachable condition in rCIP
