### 2022-05-04 9ebac66
Remove deprecated examples folder and test
