### 2022-05-04 AP6YC/ClusterValidityIndices.jl#7 @AP6YC
#19 adds an intro to the docs, cleans up the index to relevant methods, and cleans up the quickstart in both the docs and README.
