### 2022-05-04 #3527 @AP6YC
@malmaud @rMassimiliano Thank you for your patience. I've at last checked off the list of final concerns above and hope that the updates to the paper and documentation are to your liking:
1. The stable docs are back up.
2. The quickstart is trimmed up and elucidated with practical usage.
3. The full index is much better organized.
4. The paper now contains a target audience, comparison to existing methods, and CVI conceptual summary, all with additional citations.
5. Functioning DemoCard.jl examples in the documentation.
6. A code of conduct (based upon the Berlin Code of Conduct for inclusivity) is added.
7. A contributing document (that mainly points to and summarizes the extensive contributing guidelines in the docs) is added.
8. Testing coverage is now 100%, and CI is expanded to 32-bit as well as 64-bit on Windows, macOS, and Linux (Ubuntu).
9. The PkgEval (JuliaHub's CI before folding in a new version) is now green.
And more!

I am ready for you reviewers to proceed, and I am happy to make any additional change at your request to get the project and paper to your standards.
