### 2022-05-05 2499045
Manually add v0.4.0 docs
