+++ v0.4.0/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"examples/basic-examples/cvi-example.html#cvi_example","page":"CVI Simple Example","title":"CVI Simple Example","text":"","category":"section"},{"location":"examples/basic-examples/cvi-example.html","page":"CVI Simple Example","title":"CVI Simple Example","text":"(Image: Source code) (Image: notebook) (Image: compat) (Image: Author) (Image: Update time)","category":"page"},{"location":"examples/basic-examples/cvi-example.html#Overview","page":"CVI Simple Example","title":"Overview","text":"","category":"section"},{"location":"examples/basic-examples/cvi-example.html","page":"CVI Simple Example","title":"CVI Simple Example","text":"This demo is a simple example of how to use CVIs in batch mode. Here, we load a simple dataset and run a basic clustering algorithm to prescribe a set of clusters to the features. It is a combination of these features and the prescribed labels that are used to compute the criterion value. This simple example demonstrates the usage of a single CVI, but it may be substituted for any other CVI in the ClusterValidityIndices.jl package.","category":"page"},{"location":"examples/basic-examples/cvi-example.html#Clustering","page":"CVI Simple Example","title":"Clustering","text":"","category":"section"},{"location":"examples/basic-examples/cvi-example.html#Data-Setup","page":"CVI Simple Example","title":"Data Setup","text":"","category":"section"},{"location":"examples/basic-examples/cvi-example.html","page":"CVI Simple Example","title":"CVI Simple Example","text":"First, we must load all of our dependencies. We will load the ClusterValidityIndices.jl along with some data utilities and the Julia Clustering.jl package to cluster that data.","category":"page"},{"location":"examples/basic-examples/cvi-example.html","page":"CVI Simple Example","title":"CVI Simple Example","text":"using ClusterValidityIndices    # CVI/ICVI\nusing Clustering                # DBSCAN\nusing MLDatasets                # Iris dataset\nusing MLDataUtils               # Shuffling and splitting\nusing Printf                    # Formatted number printing","category":"page"},{"location":"examples/basic-examples/cvi-example.html","page":"CVI Simple Example","title":"CVI Simple Example","text":"We will download the Iris dataset for its small size and benchmark use for clustering algorithms.","category":"page"},{"location":"examples/basic-examples/cvi-example.html","page":"CVI Simple Example","title":"CVI Simple Example","text":"Iris.download(i_accept_the_terms_of_use=true)\nfeatures, labels = Iris.features(), Iris.labels()","category":"page"},{"location":"examples/basic-examples/cvi-example.html","page":"CVI Simple Example","title":"CVI Simple Example","text":"Because the MLDatasets package gives us Iris labels as strings, we will use the MLDataUtils.convertlabel method with the MLLabelUtils.LabelEnc.Indices type to get a list of integers representing each class:","category":"page"},{"location":"examples/basic-examples/cvi-example.html","page":"CVI Simple Example","title":"CVI Simple Example","text":"labels = convertlabel(LabelEnc.Indices{Int}, labels)\nunique(labels)","category":"page"},{"location":"examples/basic-examples/cvi-example.html#Fuzzy-C-Means","page":"CVI Simple Example","title":"Fuzzy C-Means","text":"","category":"section"},{"location":"examples/basic-examples/cvi-example.html","page":"CVI Simple Example","title":"CVI Simple Example","text":"Get the Fuzzy C-Means clustering result","category":"page"},{"location":"examples/basic-examples/cvi-example.html","page":"CVI Simple Example","title":"CVI Simple Example","text":"results = fuzzy_cmeans(features, 3, 2)","category":"page"},{"location":"examples/basic-examples/cvi-example.html","page":"CVI Simple Example","title":"CVI Simple Example","text":"Be
[…]
