+++ v0.4.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"examples/cvis/cvi-icvi/#cvi_icvi","page":"CVI-ICVI Comparison","title":"CVI-ICVI Comparison","text":"","category":"section"},{"location":"examples/cvis/cvi-icvi/","page":"CVI-ICVI Comparison","title":"CVI-ICVI Comparison","text":"(Image: Source code) (Image: notebook) (Image: compat) (Image: Author) (Image: Update time)","category":"page"},{"location":"examples/cvis/cvi-icvi/#Overview","page":"CVI-ICVI Comparison","title":"Overview","text":"","category":"section"},{"location":"examples/cvis/cvi-icvi/","page":"CVI-ICVI Comparison","title":"CVI-ICVI Comparison","text":"This demo is a simple example of how to use CVIs incrementally and in batch to demonstrate that the end results are the same. Here, we load a simple dataset and run a basic clustering algorithm to prescribe a set of clusters to the features. In the incremental case, we will take advantage of the fact that we can compute a criterion value at every step by running the ICVI alongside an online clustering algorithm. This simple example demonstrates the usage of a single CVI/ICVI, but it may be substituted for any other CVI/ICVI in the ClusterValidityIndices.jl package.","category":"page"},{"location":"examples/cvis/cvi-icvi/#Data-Setup","page":"CVI-ICVI Comparison","title":"Data Setup","text":"","category":"section"},{"location":"examples/cvis/cvi-icvi/","page":"CVI-ICVI Comparison","title":"CVI-ICVI Comparison","text":"First, we must load all of our dependencies. We will load the ClusterValidityIndices.jl along with some data utilities and the Julia Clustering.jl package to cluster that data.","category":"page"},{"location":"examples/cvis/cvi-icvi/","page":"CVI-ICVI Comparison","title":"CVI-ICVI Comparison","text":"using ClusterValidityIndices    # CVI/ICVI\nusing AdaptiveResonance         # DDVFA\nusing MLDatasets                # Iris dataset\nusing MLDataUtils               # Shuffling and splitting\nusing Printf                    # Formatted number printing\nusing Plots","category":"page"},{"location":"examples/cvis/cvi-icvi/","page":"CVI-ICVI Comparison","title":"CVI-ICVI Comparison","text":"We will download the Iris dataset for its small size and benchmark use for clustering algorithms.","category":"page"},{"location":"examples/cvis/cvi-icvi/","page":"CVI-ICVI Comparison","title":"CVI-ICVI Comparison","text":"Iris.download(i_accept_the_terms_of_use=true)\nfeatures, labels = Iris.features(), Iris.labels()","category":"page"},{"location":"examples/cvis/cvi-icvi/","page":"CVI-ICVI Comparison","title":"CVI-ICVI Comparison","text":"Because the MLDatasets package gives us Iris labels as strings, we will use the MLDataUtils.convertlabel method with the MLLabelUtils.LabelEnc.Indices type to get a list of integers representing each class:","category":"page"},{"location":"examples/cvis/cvi-icvi/","page":"CVI-ICVI Comparison","title":"CVI-ICVI Comparison","text":"labels = convertlabel(LabelEnc.Indices{Int}, labels)\nunique(labels)","category":"page"},{"location":"examples/cvis/cvi-icvi/#CVI/ICVI-Setup","page":"CVI-ICVI Comparison","title":"CVI/ICVI Setup","text":"","category":"section"},{"location":"examples/cvis/cvi-icvi/","page":"CVI-ICVI Comparison","title":"CVI-ICVI Comparison","text":"Because CVI/ICVIs only differ in their evaluation mode, we will use the same default constructor for 
[…]
