### 2022-07-12 bea6777
Fix package name in guide, func signatures, and typo in readme
