### 2022-07-12 #3527 @AP6YC
@malmaud Very good catch! Sorry for that oversight. I have updated it now (along with some other typos in the guide) and bumped the version of the package.
