+++ src/CVI/XB.jl
@@ -11,13 +11,13 @@ Julia port: Sasha Petrenko <sap625@mst.edu>
 REFERENCES
 [1] X. L. Xie and G. Beni, "A Validity Measure for Fuzzy Clustering," IEEE
 Transactions on Pattern Analysis and Machine Intelligence, vol. 13, no. 8,
-pp. 841–847, 1991.
+pp. 841-847, 1991.
 [2] M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey,
 "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018,
 arXiv:1801.02937v1 [stat.ML]. [Online].
 [3] M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online
 cluster validity indices for performance monitoring of streaming data clustering,"
-Int. J. Intell. Syst., pp. 1–23, 2018.
+Int. J. Intell. Syst., pp. 1-23, 2018.
 """
 
 """
@@ -26,9 +26,9 @@ Int. J. Intell. Syst., pp. 1–23, 2018.
 The stateful information of the Xie-Beni (XB) Cluster Validity Index.
 
 # References
-1. X. L. Xie and G. Beni, "A Validity Measure for Fuzzy Clustering," IEEE Transactions on Pattern Analysis and Machine Intelligence, vol. 13, no. 8, pp. 841–847, 1991.
+1. X. L. Xie and G. Beni, "A Validity Measure for Fuzzy Clustering," IEEE Transactions on Pattern Analysis and Machine Intelligence, vol. 13, no. 8, pp. 841-847, 1991.
 2. M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML]. [Online].
-3. M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1–23, 2018.
+3. M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1-23, 2018.
 """
 mutable struct XB <: CVI
     label_map::LabelMap
