+++ src/CVI/GD43.jl
@@ -11,17 +11,17 @@ Julia port: Sasha Petrenko <sap625@mst.edu>
 REFERENCES
 [1] A. Ibrahim, J. M. Keller, and J. C. Bezdek, "Evaluating Evolving Structure
 in Streaming Data With Modified Dunn's Indices," IEEE Transactions on Emerging
-Topics in Computational Intelligence, pp. 1–12, 2019.
+Topics in Computational Intelligence, pp. 1-12, 2019.
 [2] M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey,
 "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018,
 arXiv:1801.02937v1 [stat.ML].
 [3] M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online
 cluster validity indices for performance monitoring of streaming data clustering,"
-Int. J. Intell. Syst., pp. 1–23, 2018.
+Int. J. Intell. Syst., pp. 1-23, 2018.
 [4] J. C. Dunn, "A fuzzy relative of the ISODATA process and its use in detecting
-compact well-separated clusters," J. Cybern., vol. 3, no. 3 , pp. 32–57, 1973.
+compact well-separated clusters," J. Cybern., vol. 3, no. 3 , pp. 32-57, 1973.
 [5] J. C. Bezdek and N. R. Pal, "Some new indexes of cluster validity," IEEE
-Trans. Syst., Man, and Cybern., vol. 28, no. 3, pp. 301–315, Jun. 1998.
+Trans. Syst., Man, and Cybern., vol. 28, no. 3, pp. 301-315, Jun. 1998.
 
 """
 
@@ -31,11 +31,11 @@ Trans. Syst., Man, and Cybern., vol. 28, no. 3, pp. 301–315, Jun. 1998.
 The stateful information of the Generalized Dunn's Index 43 (GD43) Cluster Validity Index.
 
 # References
-1. A. Ibrahim, J. M. Keller, and J. C. Bezdek, "Evaluating Evolving Structure in Streaming Data With Modified Dunn's Indices," IEEE Transactions on Emerging Topics in Computational Intelligence, pp. 1–12, 2019.
+1. A. Ibrahim, J. M. Keller, and J. C. Bezdek, "Evaluating Evolving Structure in Streaming Data With Modified Dunn's Indices," IEEE Transactions on Emerging Topics in Computational Intelligence, pp. 1-12, 2019.
 2. M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, and J. Bailey, "Online Cluster Validity Indices for Streaming Data," ArXiv e-prints, 2018, arXiv:1801.02937v1 [stat.ML].
-3. M. Moshtaghi, J. C. Bezdek, S. M. Erfani, C. Leckie, J. Bailey, "Online cluster validity indices for performance monitoring of streaming data clustering," Int. J. Intell. Syst., pp. 1–23, 2018.
-4. J. C. Dunn, "A fuzzy relative of the ISODATA process and its use in detecting compact well-separated clusters," J. Cybern., vol. 3, no. 3 , pp. 32–57, 1973.
-5. J. C. Bezdek and N. R. Pal, "Some new indexes of cluster validity," IEEE Trans. Syst., Man, and Cybern., vol. 28, no. 3, pp. 301–315, Jun. 1998.
[…]
