### 2022-09-13 b0d6aaa
Break long equations in CH and DB
