### 2022-09-13 b20add3
Delete bloat file and ignore dev folder
