+++ src/CVI/XB.jl
@@ -89,7 +89,10 @@ function param_inc!(cvi::XB, sample::RealVector, label::Integer)
         mu_data_new = sample
         setup!(cvi, sample)
     else
-        mu_data_new = (1 - 1/n_samples_new) .* cvi.mu_data + (1/n_samples_new) .* sample
+        mu_data_new = (
+            (1 - 1 / n_samples_new) .* cvi.mu_data
+            + (1 / n_samples_new) .* sample
+        )
     end
 
     if i_label > cvi.n_clusters
@@ -105,8 +108,7 @@ function param_inc!(cvi::XB, sample::RealVector, label::Integer)
             d_column_new = zeros(cvi.n_clusters + 1)
             # println(d_column_new)
             for jx = 1:cvi.n_clusters
-                # sum((v_new - cvi.v[:, jx]).^2, dims=1)
-                d_column_new[jx] = sum((v_new - cvi.v[:, jx]).^2)
+                d_column_new[jx] = sum((v_new - cvi.v[:, jx]) .^ 2)
             end
             D_new[:, i_label] = d_column_new
             D_new[i_label, :] = transpose(d_column_new)
@@ -121,17 +123,24 @@ function param_inc!(cvi::XB, sample::RealVector, label::Integer)
         cvi.D = D_new
     else
         n_new = cvi.n[i_label] + 1
-        v_new = (1 - 1/n_new) .* cvi.v[:, i_label] + (1/n_new) .* sample
+        v_new = (
+            (1 - 1/n_new) .* cvi.v[:, i_label]
+            + (1/n_new) .* sample
+        )
         delta_v = cvi.v[:, i_label] - v_new
         diff_x_v = sample .- v_new
-        CP_new = cvi.CP[i_label] + transpose(diff_x_v)*diff_x_v + cvi.n[i_label]*transpose(delta_v)*delta_v + 2*transpose(delta_v)*cvi.G[:, i_label]
+        CP_new = (
+            cvi.CP[i_label]
+            + transpose(diff_x_v) * diff_x_v
+            + cvi.n[i_label] * transpose(delta_v) * delta_v
+            + 2 * transpose(delta_v) * cvi.G[:, i_label]
+        )
         d_column_new = zeros(cvi.n_clusters)
         for jx = 1:cvi.n_clusters
             if jx == i_label
                 continue
             end
-            # sum((v_new - cvi.v[:, jx]).^2, dims=1)
-            d_column_new[jx] = sum((v_new - cvi.v[:, jx]).^2)
+            d_column_new[jx] = sum((v_new - cvi.v[:, jx]) .^ 2)
         end
         # Update parameters
         cvi.n[i_label] = n_new
@@ -160,12 +169,13 @@ function param_batch!(cvi::XB, data::RealMatrix, labels::IntegerVector)
         cvi.n[ix] = size(subset, 2)
         cvi.v[1:cvi.dim, ix] = mean(subset, dims=2)
         diff_x_v = subset - cvi.v[:, ix] * ones(1, cvi.n[ix])
-        cvi.CP[ix] = sum(diff_x_v.^2)
+        cvi.CP[ix] = sum(diff_x_v .^ 2)
     end
     for ix = 1 : (cvi.n_clusters - 1)
         for jx = ix + 1 : cvi.n_clusters
-            # cvi.D[jx, ix] = sum((cvi.v[:, ix] - cvi.v[:, jx]).^2, dims=1)
-            cvi.D[jx, ix] = sum((cvi.v[:, ix] - cvi.v[:, jx]).^2)
+            cvi.D[jx, ix] = (
+                sum((cvi.v[:, ix] - cvi.v[:, jx]) .^ 2)
+            )
         end
     end
[…]
