+++ src/CVI/cSIL.jl
@@ -111,8 +111,8 @@ function param_inc!(cvi::cSIL, sample::RealVector, label::Integer)
                 # Row "bmu_temp" - E
                 C = (
                     cvi.CP[cl]
-                    + cvi.n[cl] * (transpose(v_new)*v_new)
-                    - 2*(transpose(cvi.G[:, cl]) * v_new)
+                    + cvi.n[cl] * (transpose(v_new) * v_new)
+                    - 2 * (transpose(cvi.G[:, cl]) * v_new)
                 )
                 S_row_new[cl] = C / cvi.n[cl]
             end
@@ -133,8 +133,8 @@ function param_inc!(cvi::cSIL, sample::RealVector, label::Integer)
     else
         n_new = cvi.n[i_label] + 1
         v_new = (
-            (1 - 1/n_new) .* cvi.v[:, i_label]
-            + (1/n_new) .* sample
+            (1 - 1 / n_new) .* cvi.v[:, i_label]
+            + (1 / n_new) .* sample
         )
         CP_new = cvi.CP[i_label] + (transpose(sample) * sample)
         G_new = cvi.G[:, i_label] + sample
@@ -150,8 +150,8 @@ function param_inc!(cvi::cSIL, sample::RealVector, label::Integer)
             diff_x_v = sample - cvi.v[:, cl]
             C = (
                 cvi.CP[i_label]
-                + (transpose(diff_x_v)*diff_x_v)
-                + cvi.n[i_label]*(transpose(cvi.v[:, cl]) * cvi.v[:, cl])
+                + (transpose(diff_x_v) * diff_x_v)
+                + cvi.n[i_label] * (transpose(cvi.v[:, cl]) * cvi.v[:, cl])
                 - 2 * (transpose(G_new) * cvi.v[:, cl])
             )
             S_col_new[cl] = C / n_new
@@ -169,7 +169,7 @@ function param_inc!(cvi::cSIL, sample::RealVector, label::Integer)
             cvi.CP[i_label]
             + (transpose(diff_x_v) * diff_x_v)
             + cvi.n[i_label] * (transpose(v_new) * v_new)
-            - 2*(transpose(cvi.G[:, i_label])*v_new)
+            - 2 * (transpose(cvi.G[:, i_label]) * v_new)
         )
         S_col_new[i_label] = C / n_new
         S_row_new[i_label] = S_col_new[i_label]
@@ -199,7 +199,7 @@ function param_batch!(cvi::cSIL, data::RealMatrix, labels::IntegerVector)
         cvi.n[ix] = size(subset, 2)
         cvi.v[:, ix] = mean(subset, dims=2)
         # Compute CP in case of switching back to incremental mode
-        d_temp = (data - cvi.v[:, ix]*ones(1, cvi.n_samples)).^2
+        d_temp = (data - cvi.v[:, ix] * ones(1, cvi.n_samples)) .^ 2
         D[:, ix] = transpose(sum(d_temp, dims=1))
     end
     for ix = 1:cvi.n_clusters
@@ -218,7 +218,7 @@ function evaluate!(cvi::cSIL)
             a = cvi.S[ix, ix]
             # Other clusters
             b = minimum(cvi.S[ix, 1:end .!= ix])
-            cvi.sil_coefs[ix] = (b-a)/max(a, b)
+            cvi.sil_coefs[ix] = (b - a) / max(a, b)
         end
         # cSIL index value
         cvi.criterion_value = sum(cvi.sil_coefs) / cvi.n_clusters

[…]
