+++ paper/paper.bib
@@ -1,0 +1,104 @@
+@article{ARBELAITZ2013243,
+    title = {An extensive comparative study of cluster validity indices},
+    journal = {Pattern Recognition},
+    volume = {46},
+    number = {1},
+    pages = {243-256},
+    year = {2013},
+    issn = {0031-3203},
+    doi = {https://doi.org/10.1016/j.patcog.2012.07.021},
+    url = {https://www.sciencedirect.com/science/article/pii/S003132031200338X},
+    author = {Olatz Arbelaitz and Ibai Gurrutxaga and Javier Muguerza and Jesús M. Pérez and Iñigo Perona},
+    keywords = {Crisp clustering, Cluster validity index, Comparative analysis},
+    abstract = {The validation of the results obtained by clustering algorithms is a fundamental part of the clustering process. The most used approaches for cluster validation are based on internal cluster validity indices. Although many indices have been proposed, there is no recent extensive comparative study of their performance. In this paper we show the results of an experimental work that compares 30 cluster validity indices in many different environments with different characteristics. These results can serve as a guideline for selecting the most suitable index for each possible application and provide a deep insight into the performance differences between the currently available indices.}
+}
+
+@inproceedings{CVI_Survey_Garcia,
+    author = {Jos\'{e}-Garc\'{\i}a, Ad\'{a}n and G\'{o}mez-Flores, Wilfrido},
+    title = {A Survey of Cluster Validity Indices for Automatic Data Clustering Using Differential Evolution},
+    year = {2021},
+    isbn = {9781450383509},
+    publisher = {Association for Computing Machinery},
+    address = {New York, NY, USA},
+    url = {https://doi.org/10.1145/3449639.3459341},
+    doi = {10.1145/3449639.3459341},
+    abstract = {In cluster analysis, the automatic clustering problem refers to the determination of both the appropriate number of clusters and the corresponding natural partitioning. This can be addressed as an optimization problem in which a cluster validity index (CVI) is used as a fitness function to evaluate the quality of potential solutions. Different CVIs have been proposed in the literature, aiming to identify adequate cluster solutions in terms of intracluster cohesion and intercluster separation. However, it is important to identify the scenarios in which these CVIs perform well and their limitations. This paper evaluates the effectiveness of 22 different CVIs used as fitness functions in an evolutionary clustering algorithm named ACDE based on differential evolution. Several synthetic datasets are considered: linearly separable data having both well-separated and overlapped clusters, and non-linearly separable data having arbitrarily-shaped clusters. Besides, real-life datasets are also considered. The experimental results indicate that the Silhouette index consistently reached an acceptable performance in linearly separable data. Furthermore, the indices Calinski-Harabasz, Davies-Bouldin, and generalized Dunn obtained an adequate clustering performance in synthetic and real-life datasets. Notably, all the evaluated CVIs performed poorly in clustering the non-linearly separable data because of the assumptions about data distributions.},
+    booktitle = {Proceedings of the Genetic and Evolutionary Computation Conference},
+    pages = {314-322},
+    numpages = {9},
+    keywords = {cluster validity index, automatic clustering, differential evolution},
+    location = {Lille, France},
+    series = {GECCO '21}
+}
+
+@Manual{R_dtwclust,
+    title = {Time Series Clustering Along with Optimizations for the Dynamic Time Warping Distance},
+    author = {Alexis Sarda-Espinosa},
+    year = {2022},
[…]
