### 2022-09-13 07e4252
Remove JOSS paper and workflow
