### 2022-09-15 AP6YC/ClusterValidityIndices.jl#29 @AP6YC
Code idiosyncracies
The CVI code has some extra computations in some sections with no purpose. Trimming these lines can improve computational performance.

Furthermore, the README documentation contains some broken and deprecated links that should be updated.
