+++ test/test_cvis.jl
@@ -26,6 +26,7 @@
 
     # Sanitize all the data
     p = 0.1
+    # p = 1
     @info "p is a $(typeof(p))"
 
     @info "Subsampling data at rate: $p"
@@ -89,6 +90,7 @@
                 _ = get_icvi!(cvi, sample, label)
                 # cvs_ip[data_path][ix, cx] = cv
             end
+            @info "CVI: $(typeof(cvi)), index: $(cvi.criterion_value)"
         end
     end
 
