+++ src/CVI/GD53.jl
@@ -193,8 +193,8 @@ function param_batch!(cvi::GD53, data::RealMatrix, labels::IntegerVector)
 end # param_batch!(cvi::GD53, data::RealMatrix, labels::IntegerVector)
 
 function evaluate!(cvi::GD53)
-    cvi.intra = 2*maximum(cvi.CP ./ cvi.n)
     if cvi.n_clusters > 1
+        cvi.intra = 2 * maximum(cvi.CP ./ cvi.n)
         # Between-group measure of separation/isolation
         cvi.inter = (
             minimum(cvi.D[triu(ones(Bool, cvi.n_clusters, cvi.n_clusters), 1)])
