### 2022-09-20 6e8716e
Remove plotting test reqs, add Printf dep
