### 2022-09-29 a0db478
Load local package in docs
