### 2022-10-03 #3527 @rMassimiliano
@adi3 Sorry that it took more than expected.

@AP6YC can you confirm that you addressed my previous comments?

By looking at the package documentation and at the Readme it does not seem so.

###### Examples:

1. the software is still under documented. The documentation of the function `DB()` - that is used as example - has not been updated. It still read:
```
 DB()

  Default constructor for the Davies-Bouldin (DB) Cluster Validity Index.
```

2. The `README.md` file refers to an Examples section (https://ap6yc.github.io/ClusterValidityIndices.jl/dev/man/examples/) that does not exist. And no executable example is provided in the `README.md`.
