### 2022-10-03 cc3d1fa
Remove results and examples folders in readme
