### 2022-10-03 dbffe87
Update examples section of readme, correct links
