### 2022-10-03 #3527 @AP6YC
@rMassimiliano Thank you for your feedback! I can immediately address both of your points:

1. I chose to document the citations and information about the CVIs in the structs themselves, listed under the [full-index](https://ap6yc.github.io/ClusterValidityIndices.jl/dev/man/full-index/). However, I agree that some more information could be included in the constructor docstrings themselves.
2. I am fixing those links now in the documentation in addition to some more idiosyncrasies that I have found throughout. I will also update the README guide to contain executable code.

Both of these documentation-oriented fixes will be under solved in [issue 29](https://github.com/AP6YC/ClusterValidityIndices.jl/issues/29). I will provide an update when I fold in these changes for your review.
