### 2022-10-04 e2aca31
Add docstring extensions attributes, template
