+++ src/common.jl
@@ -5,9 +5,23 @@ Description:
     All common types, aliases, structs, and methods for the ClusterValidityIndices.jl package.
 """
 
-# -------------------------------------------
-# Abstract types
-# -------------------------------------------
+# --------------------------------------------------------------------------- #
+# DOCSTRING TEMPLATES
+# --------------------------------------------------------------------------- #
+
+@template TYPES =
+"""
+$(TYPEDEF)
+
+$(DOCSTRING)
+
+# Typed Fields
+$(TYPEDFIELDS)
+"""
+
+# --------------------------------------------------------------------------- #
+# ABSTRACT TYPES
+# --------------------------------------------------------------------------- #
 
 # Type for all CVIs
 """
