+++ docs/Project.toml
@@ -10,6 +10,7 @@ LiveServer = "16fef848-5104-11e9-1b77-fb7a48bbb589"
 MLDataUtils = "cc2ba9b6-d476-5e6d-8eaf-a92d5412d41d"
 MLDatasets = "eb30cadb-4394-5ae3-aed4-317e484a6458"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 
 [compat]
 MLDatasets = "0.7"
