### 2022-10-04 fb07ea8
Dev templates, trim existing docstrings
