+++ docs/src/man/dev-index.md
@@ -1,0 +1,33 @@
+# [Index](@id dev-main-index)
+
+This page lists the core methods and types of the `ClusterValidityIndices.jl` package.
+The [Methods](@ref index-methods) section lists the public methods for the package that use the CVIs/ICVIs in [Types](@ref index-types).
+Each of these entries link to the docstrings in the [Docs](@ref index-docs) section.
+
+## [Methods](@id dev-index-methods)
+
+```@index
+Pages   = ["man/dev-index.md"]
+Modules = [ClusterValidityIndices]
+Order = [:function]
+```
+
+## [Types](@id dev-index-types)
+
+```@index
+Pages   = ["man/dev-index.md"]
+Modules = [ClusterValidityIndices]
+Order = [:type]
+```
+
+## [Docs](@id dev-index-docs)
+
+<!-- ```@autodocs
+Modules = [ClusterValidityIndices]
+Private = true
+``` -->
+
+```@docs
+ClusterValidityIndices.setup!
+ClusterValidityIndices.get_internal_label!
+```
