+++ docs/src/index.md
@@ -25,6 +25,7 @@ Pages = [
     "../examples/index.md",
     "man/contributing.md",
     "man/full-index.md",
+    "man/dev-index.md",
 ]
 Depth = 1
 ```
