### 2022-10-06 bf3aa69
Rework full index, dev internals
