+++ docs/src/man/full-index.md
@@ -4,41 +4,38 @@ This page lists the core methods and types of the `ClusterValidityIndices.jl` pa
 The [Methods](@ref index-methods) section lists the public methods for the package that use the CVIs/ICVIs in [Types](@ref index-types).
 Each of these entries link to the docstrings in the [Docs](@ref index-docs) section.
 
-## [Methods](@id index-methods)
+## Index
+
+### [Functions](@id index-methods)
 
 ```@index
+Pages   = ["full-index.md"]
 Modules = [ClusterValidityIndices]
 Order = [:function]
-Public = true
 ```
 
-## [Types](@id index-types)
+### [Types](@id index-types)
 
 ```@index
+Pages   = ["full-index.md"]
 Modules = [ClusterValidityIndices]
 Order = [:type]
-Public = true
 ```
 
 ## [Docs](@id index-docs)
 
-```@docs
-ClusterValidityIndices
-get_cvi!
-get_icvi!
-param_inc!
-param_batch!
-evaluate!
-relabel_cvi_data
-sort_cvi_data
-CH
-cSIL
-DB
-GD43
-GD53
-PS
-rCIP
-WB
-XB
-CVI
+### Functions
+
+```@autodocs
+Modules = [ClusterValidityIndices]
+Private = false
+Order = [:function]
+```
+
+### Types
+
+```@autodocs
+Modules = [ClusterValidityIndices]
+Private = false
+Order = [:type]
 ```
