+++ docs/src/man/full-index.md
@@ -1,10 +1,19 @@
 # [Index](@id main-index)
 
 This page lists the core methods and types of the `ClusterValidityIndices.jl` package.
+The [Modules](@ref index-modules) section lists the modules exported by the package including the
 The [Methods](@ref index-methods) section lists the public methods for the package that use the CVIs/ICVIs in [Types](@ref index-types).
 Each of these entries link to the docstrings in the [Docs](@ref index-docs) section.
 
 ## Index
+
+### [Modules](@id index-modules)
+
+```@index
+Pages   = ["full-index.md"]
+Modules = [ClusterValidityIndices]
+Order = [:module]
+```
 
 ### [Functions](@id index-methods)
 
@@ -23,6 +32,14 @@ Order = [:type]
 ```
 
 ## [Docs](@id index-docs)
+
+### Modules
+
+```@autodocs
+Modules = [ClusterValidityIndices]
+Private = false
+Order = [:module]
+```
 
 ### Functions
 

[…]
