### 2022-10-06 879a720
Add references docstring vars
