+++ src/common.jl
@@ -155,6 +155,10 @@ end # get_cvi!(cvi::CVI, data::RealMatrix, labels::IntegerVector)
 
 """
 Get the internal label and update the label map if the label is new.
+
+# Arguments
+- `label_map::LabelMap`: label map to extract the internal label from.
+- `label::Int`: the external label that corresponds to an internal label.
 """
 function get_internal_label!(label_map::LabelMap, label::Int)
     # If the label map contains the key, return that internal label
@@ -237,8 +241,10 @@ julia> my_criterion_value = my_cvi.criterion_value
 evaluate!(cvi::CVI)
 
 @doc raw"""
-    setup!(cvi::CVI, sample::Vector{T}) where {T<:RealFP}
-
 Internal method, sets up the CVI based upon the type of the provided sample.
+
+# Arguments
+- `cvi::CVI`: the CVI to setup to the correct dimensions.
+- `sample::Vector{T<:RealFP}`: The sample to use as a basis for setting up the CVI.
 """
 setup!(cvi::CVI, sample::Vector{T}) where {T<:RealFP}
