+++ docs/src/man/full-index.md
@@ -1,11 +1,13 @@
 # [Index](@id main-index)
 
 This page lists the core methods and types of the `ClusterValidityIndices.jl` package.
-The [Modules](@ref index-modules) section lists the modules exported by the package including the
+The [Modules](@ref index-modules) section lists the modules exported by the package including the `ClusterValidityIndices` module itself.
 The [Methods](@ref index-methods) section lists the public methods for the package that use the CVIs/ICVIs in [Types](@ref index-types).
 Each of these entries link to the docstrings in the [Docs](@ref index-docs) section.
 
 ## Index
+
+This section enumerates the names exported by the package, each of which links to its corresponding [Documentation](@ref index-docs).
 
 ### [Modules](@id index-modules)
 
@@ -32,6 +34,8 @@ Order = [:type]
 ```
 
 ## [Docs](@id index-docs)
+
+This section lists the documentation for every exported name of the `ClusterValidityIndices.jl` package.
 
 ### Modules
 
