+++ docs/make.jl
@@ -12,13 +12,14 @@ using
 
 # Get the current workind directory's base name
 current_dir = basename(pwd())
+@info "Current directory is $(current_dir)"
 
 # If using the CI method `julia --project=docs/ docs/make.jl`
 #   or `julia --startup-file=no --project=docs/ docs/make.jl`
-if current_dir == "ClusterValidityIndices"
+if occursin("ClusterValidityIndices", current_dir)
     push!(LOAD_PATH, "../src/")
 # Otherwise, we are already in the docs project and need to dev the above package
-elseif current_dir == "docs"
+elseif occursin("docs", current_dir)
     Pkg.develop(path="..")
 # Otherwise, building docs from the wrong path
 else
