+++ src/common.jl
@@ -48,33 +48,9 @@ All index instantiations are subtypes of `CVI`.
 """
 abstract type CVI end
 
-# -------------------------------------------
-# Aliases
-# -------------------------------------------
-#   **Taken from StatsBase.jl**
-#
-#  These types signficantly reduces the need of using
-#  type parameters in functions (which are often just
-#  for the purpose of restricting the arrays to real)
-#
-# These could be removed when the Base supports
-# covariant type notation, i.e. AbstractVector{<:Real}
-
-# Real-numbered aliases
-const RealArray{T<:Real, N} = AbstractArray{T, N}
-const RealVector{T<:Real} = AbstractArray{T, 1}
-const RealMatrix{T<:Real} = AbstractArray{T, 2}
-
-# Integered aliases
-const IntegerArray{T<:Integer, N} = AbstractArray{T, N}
-const IntegerVector{T<:Integer} = AbstractArray{T, 1}
-const IntegerMatrix{T<:Integer} = AbstractArray{T, 2}
-
-# Specifically floating-point aliases
-const RealFP = Union{Float32, Float64}
-
-# System's largest native floating point variable
-const Float = (Sys.WORD_SIZE == 64 ? Float64 : Float32)
+# --------------------------------------------------------------------------- #
+# CONSTANTS
+# --------------------------------------------------------------------------- #
 
 """
 Internal label mapping for incremental CVIs.
@@ -84,7 +60,7 @@ Alias for a dictionary mapping of integers to integers as cluster labels.
 const LabelMap = Dict{Int, Int}
 
 # --------------------------------------------------------------------------- #
-# Methods
+# FUNCTIONS
 # --------------------------------------------------------------------------- #
 
 """
@@ -175,9 +151,9 @@ function get_internal_label!(label_map::LabelMap, label::Int)
     return internal_label
 end # get_internal_label!(label_map::LabelMap, label::Int)
 
-# -------------------------------------------
-# Common Documentation
-# -------------------------------------------
+# --------------------------------------------------------------------------- #
+# COMMON DOCUMENTATION
+# --------------------------------------------------------------------------- #
 
 @doc raw"""
 Compute the CVI parameters incrementally.
