### 2022-10-06 551234c
Remove unused names, add test file docstrings
