### 2022-10-06 fc0b79c
Replace String to AbstractString where applicable
