+++ src/CVI/CH.jl
@@ -39,17 +39,17 @@ $(local_references)
 """
 mutable struct CH <: CVI
     label_map::LabelMap
-    dim::Integer
-    n_samples::Integer
-    mu::RealVector      # dim
-    n::IntegerVector    # dim
-    v::RealMatrix       # dim x n_clusters
-    CP::RealVector      # dim
-    SEP::RealVector     # dim
-    G::RealMatrix       # dim x n_clusters
+    dim::Int
+    n_samples::Int
+    mu::Vector{Float}   # dim
+    n::Vector{Int}      # dim
+    v::Matrix{Float}    # dim x n_clusters
+    CP::Vector{Float}   # dim
+    SEP::Vector{Float}  # dim
+    G::Matrix{Float}    # dim x n_clusters
     BGSS::Float
     WGSS::Float
-    n_clusters::Integer
+    n_clusters::Int
     criterion_value::Float
 end # CH <: CVI
 
@@ -63,12 +63,12 @@ function CH()
         LabelMap(),                     # label_map
         0,                              # dim
         0,                              # n_samples
-        Array{Float, 1}(undef, 0),      # mu
-        Array{Integer, 1}(undef, 0),    # n
-        Array{Float, 2}(undef, 0, 0),   # v
-        Array{Float, 1}(undef, 0),      # CP
-        Array{Float, 1}(undef, 0),      # SEP
-        Array{Float, 2}(undef, 0, 0),   # G
+        Vector{Float}(undef, 0),        # mu
+        Vector{Int}(undef, 0),          # n
+        Matrix{Float}(undef, 0, 0),     # v
+        Vector{Float}(undef, 0),        # CP
+        Vector{Float}(undef, 0),        # SEP
+        Matrix{Float}(undef, 0, 0),     # G
         0.0,                            # BGSS
         0.0,                            # WGSS
         0,                              # n_clusters
@@ -76,15 +76,17 @@ function CH()
     )
 end # CH()
 
-function setup!(cvi::CH, sample::Vector{T}) where {T<:RealFP}
+# Setup function
+function setup!(cvi::CH, sample::RealVector)
     # Get the feature dimension
     cvi.dim = length(sample)
     # Initialize the augmenting 2-D arrays with the correct feature dimension
     # NOTE: R is emptied and calculated in evaluate!, so it is not defined here
-    cvi.v = Array{T, 2}(undef, cvi.dim, 0)
-    cvi.G = Array{T, 2}(undef, cvi.dim, 0)
+    cvi.v = Matrix{Float}(undef, cvi.dim, 0)
+    cvi.G = Matrix{Float}(undef, cvi.dim, 0)
 end # setup!(cvi::CH, sample::Vector{T}) where {T<:Real}
 
+# Incremental parameter update function
 function param_inc!(cvi::CH, sample::RealVector, label::Integer)
     # Get the internal label
     i_label = get_internal_label!(cvi.label_map, label)
@@ -142,6 +144,7 @@ function param_inc!(cvi::CH, sample::RealVector, label::Integer)
     )
 end # param_inc!(cvi::CH, sample::RealVector, label::Integer)
 
+# Batch parameter update function
 function param_batch!(cvi::CH, data::RealMatrix, labels::IntegerVector)
     cvi.dim, cvi.n_samples = size(data)
     # Take the average across all samples, but cast to 1-D vector
@@ -163,6 +166,7 @@ function param_batch!(cvi::CH, data::RealMatrix, labels::IntegerVector)
     end
 end # param_batch!(cvi::CH, data::RealMatrix, labels::IntegerVector)
 
+# Criterion value evaluation function
 function evaluate!(cvi::CH)
     # Within group sum of scatters
     cvi.WGSS = sum(cvi.CP)
