### 2022-10-07 5ff0f86
DB type cleanup
