+++ src/CVI/DB.jl
@@ -36,17 +36,17 @@ $(local_references)
 """
 mutable struct DB <: CVI
     label_map::LabelMap
-    dim::Integer
-    n_samples::Integer
-    mu_data::RealVector         # dim
-    n::IntegerVector            # dim
-    v::RealMatrix               # dim x n_clusters
-    CP::RealVector              # dim
-    S::RealVector               # dim
-    R::RealMatrix               # dim x n_clusters
-    G::RealMatrix               # dim x n_clusters
-    D::RealMatrix               # n_clusters x n_clusters
-    n_clusters::Integer
+    dim::Int
+    n_samples::Int
+    mu_data::Vector{Float}         # dim
+    n::Vector{Int}                 # dim
+    v::Matrix{Float}               # dim x n_clusters
+    CP::Vector{Float}              # dim
+    S::Vector{Float}               # dim
+    R::Matrix{Float}               # dim x n_clusters
+    G::Matrix{Float}               # dim x n_clusters
+    D::Matrix{Float}               # n_clusters x n_clusters
+    n_clusters::Int
     criterion_value::Float
 end # DB <: CVI
 
@@ -60,29 +60,31 @@ function DB()
         LabelMap(),                     # label_map
         0,                              # dim
         0,                              # n_samples
-        Array{Float, 1}(undef, 0),      # mu_data
-        Array{Integer, 1}(undef, 0),    # n
-        Array{Float, 2}(undef, 0, 0),   # v
-        Array{Float, 1}(undef, 0),      # CP
-        Array{Float, 1}(undef, 0),      # S
-        Array{Float, 2}(undef, 0, 0),   # R
-        Array{Float, 2}(undef, 0, 0),   # G
-        Array{Float, 2}(undef, 0, 0),   # D
+        Vector{Float}(undef, 0),        # mu_data
+        Vector{Int}(undef, 0),          # n
+        Matrix{Float}(undef, 0, 0),     # v
+        Vector{Float}(undef, 0),        # CP
+        Vector{Float}(undef, 0),        # S
+        Matrix{Float}(undef, 0, 0),     # R
+        Matrix{Float}(undef, 0, 0),     # G
+        Matrix{Float}(undef, 0, 0),     # D
         0,                              # n_clusters
         0.0                             # criterion_value
     )
 end # DB()
 
-function setup!(cvi::DB, sample::Vector{T}) where {T<:RealFP}
+# Setup function
+function setup!(cvi::DB, sample::RealVector)
     # Get the feature dimension
     cvi.dim = length(sample)
     cvi.mu_data = sample
     # Initialize the augmenting 2-D arrays with the correct feature dimension
     # NOTE: R is emptied and calculated in evaluate!, so it is not defined here
-    cvi.v = Array{T, 2}(undef, cvi.dim, 0)
-    cvi.G = Array{T, 2}(undef, cvi.dim, 0)
-end # setup!(cvi::DB, sample::Vector{T}) where {T<:RealFP}
+    cvi.v = Matrix{Float}(undef, cvi.dim, 0)
+    cvi.G = Matrix{Float}(undef, cvi.dim, 0)
+end # setup!(cvi::DB, sample::RealVector)
 
+# Incremental parameter update function
 function param_inc!(cvi::DB, sample::RealVector, label::Integer)
     # Get the internal label
     i_label = get_internal_label!(cvi.label_map, label)
@@ -164,6 +166,7 @@ function param_inc!(cvi::DB, sample::RealVector, label::Integer)
     cvi.n_samples = n_samples_new
 end # param_inc!(cvi::DB, sample::RealVector, label::Integer)
 
+# Batch parameter update function
[…]
