### 2022-10-07 eb96eb7
GD43 types cleanup
