### 2022-10-07 1e85611
GD53 types cleanup
