### 2022-10-07 bf82b37
rCIP type cleanup, formatting
