### 2022-10-07 0c9b689
XB type cleanup, formatting
