### 2022-10-07 a806b82
Clean test types
