### 2022-10-07 2f58c93
Fix common and util docstrings
