### 2022-10-07 13b4a54
CompatHelper: add new compat entry for Revise at version 3, (keep existing compat)
