+++ Project.toml
@@ -17,6 +17,7 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 DocStringExtensions = "0.8, 0.9"
 NumericalTypeAliases = "0.1"
 ProgressBars = "0.7, 0.8, 1"
+Revise = "3"
 julia = "1"
 
 [extras]
