### 2022-10-07 0915686
Create cvi constructor function
