+++ test/test_cvis.jl
@@ -8,12 +8,10 @@ A single test set for the testing the functionality of all CVIS modules.
 - Sasha Petrenko <sap625@mst.edu>
 """
 
-@testset "CVIs" begin
-    @info "CVI Testing"
-
-    # Set the approximation CVI tolerance for all comparisons
-    tolerance = 1e-1
-
+"""
+Constructs and returns a list of all cvis
+"""
+function construct_cvis()
     # Construct the cvis
     cvis = [
         CH(),
@@ -26,7 +24,14 @@ A single test set for the testing the functionality of all CVIS modules.
         WB(),
         XB(),
     ]
-    n_cvis = length(cvis)
+    return cvis
+end
+
+@testset "CVIs" begin
+    @info "CVI Testing"
+
+    # Set the approximation CVI tolerance for all comparisons
+    tolerance = 1e-1
 
     # Grab all the data paths for testing
     data_paths = readdir("../data", join=true)
@@ -55,7 +60,7 @@ A single test set for the testing the functionality of all CVIS modules.
     cvi_ip = Dict()
     for data_path in data_paths
         @info "Data: $data_path"
-        cvi_ip[data_path] = deepcopy(cvis)
+        cvi_ip[data_path] = construct_cvis()
         for cvi in cvi_ip[data_path]
             for ix = 1:n_samples[data_path]
                 sample = data[data_path][:, ix]
@@ -71,19 +76,15 @@ A single test set for the testing the functionality of all CVIS modules.
     cvi_bp = Dict()
     for data_path in data_paths
         @info "Data: $data_path"
-        cvi_bp[data_path] = deepcopy(cvis)
-        # cvs_b = zeros(n_cvis)
-        # for cx = 1:n_cvis
+        cvi_bp[data_path] = construct_cvis()
         for cvi in cvi_bp[data_path]
-            # cvs_b[cx] = get_cvi!(cvi_bp[cx], data, labels)
-            # _ = get_cvi!(cvi_bp[data_path][cx], data[data_path], labels[data_path])
             _ = get_cvi!(cvi, data[data_path], labels[data_path])
         end
     end
 
     # Test that all permutations are equivalent
     for data_path in data_paths
-        for cx = 1:n_cvis
+        for cx in eachindex(cvi_ip[data_path])
             # IP to BP
             @test isapprox(cvi_ip[data_path][cx].criterion_value,
                 cvi_bp[data_path][cx].criterion_value,
