+++ Project.toml
@@ -10,7 +10,6 @@ Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 NumericalTypeAliases = "be9b823e-291e-41a1-b8ce-806204e78f92"
 ProgressBars = "49802e3a-d2f1-5c88-81d8-b72133a6f568"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
