### 2022-10-07 281616c
Make internal variable CVI_MODULES, document
