### 2022-10-07 AP6YC/ClusterValidityIndices.jl#29 @AP6YC
Closing due to #30
