### 2022-10-07 32d471e
Add rCIP edge case test
