### 2022-10-07 54c339f
GD43 dots
