### 2022-10-07 7f467aa
Add constants docstring template
