### 2022-10-07 bd82ad7
Remove comment template
