+++ src/common.jl
@@ -9,13 +9,6 @@ Description:
 # DOCSTRING TEMPLATES
 # --------------------------------------------------------------------------- #
 
-# Default template if not specified (i.e., constants and modules)
-# @template DEFAULT =
-# """
-# $(SIGNATURES)
-# $(DOCSTRING)
-# """
-
 # Constants template
 @template CONSTANTS =
 """
