### 2022-10-11 e217dc6
Remove duplicate signatures in doc template
