### 2022-10-11 4e14f34
Remove get_icvi usage
