### 2022-10-11 be80a2c
CompatHelper: bump compat for NumericalTypeAliases to 0.2, (keep existing compat)
