### 2022-10-11 d52f1e4
Massively simplify readme
