+++ README.md
@@ -55,13 +55,13 @@ Please read the [documentation](https://ap6yc.github.io/ClusterValidityIndices.j
   - [Overview](#overview)
   - [Installation](#installation)
   - [Quickstart](#quickstart)
-    - [Examples](#examples)
   - [Implemented CVI/ICVIs](#implemented-cviicvis)
-  - [Usage](#usage)
+    - [Examples](#examples)
+  - [Detailed Usage](#detailed-usage)
     - [Data](#data)
     - [Instantiation](#instantiation)
     - [Incremental vs. Batch](#incremental-vs-batch)
-    - [Updating](#updating)
+    - [Advanced Usage](#advanced-usage)
     - [Criterion Values](#criterion-values)
     - [Porcelain](#porcelain)
   - [Structure](#structure)
@@ -76,7 +76,7 @@ Cluster Validity Indices (CVIs) are designed to be metrics of performance for un
 In the absense of supervisory labels (i.e., ground truth), clustering algorithms - or any truly unsupervised learning algorithms - have no way to definitively know the stability of their learning and accuracy of their performance.
 As a result, CVIs exist to provide metrics of partitioning stability/validity through the use of only the original data samples and the cluster labels prescribed by the clustering algorithm.
 
-This Julia package contains an outline of the conceptual usage of CVIs along with many example scripts.
+This Julia package contains an outline of the conceptual usage of CVIs along with many [example scripts in the documentation](https://ap6yc.github.io/ClusterValidityIndices.jl/dev/examples/).
 This outline begins with [a list of CVIs](#implemented-cviicvis) that are implemented in the lastest version of the project.
 [Quickstart](#quickstart) provides an overview of how to use this project, while [Structure](#structure) outlines the project file structure, giving context to the locations of every component of the project.
 [Usage](#usage) outlines the general syntax and workflow of the CVIs/ICVIs.
@@ -86,21 +86,23 @@ This outline begins with [a list of CVIs](#implemented-cviicvis) that are implem
 This project is distributed as a Julia package, available on [JuliaHub](https://juliahub.com/).
 Its usage follows the usual Julia package installation procedure, interactively:
 
-```julia
-] add ClusterValidityIndices
+```julia-repl
+julia> ]
+(@v1.8) pkg> add ClusterValidityIndices
 ```
 
 or programmatically:
 
-```julia
-using Pkg
-Pkg.add("ClusterValidityIndices")
+```julia-repl
+julia> using Pkg
+julia> Pkg.add("ClusterValidityIndices")
 ```
 
 You may also add the package directly from GitHub to get the latest changes between releases:
 
-```julia
-] add https://github.com/AP6YC/ClusterValidityIndices.jl
+```julia-repl
+julia> ]
+(@v1.8) pkg> add https://github.com/AP6YC/ClusterValidityIndices.jl
 ```
 
 ## Quickstart
@@ -108,43 +110,60 @@ You may also add the package directly from GitHub to get the latest changes betw
 This section provides a quick overview of how to use the project.
 For more detailed code usage, please see [Usage](#usage).
 
-First, assume that you have a dataset of features/data and labels prescribed by some clustering algorithm:
+First, import the package with:
 
 ```julia
-data_file = "path/to/data.csv"
-data, labels = get_cvi_data(data_file)
+# Import the package
+using ClusterValidityIndices
 ```
 
-All CVIs in this package are acronymed (see [Implemented Modules](#implemented-cviicvis)).
-You can create a new CVI structure with a default constructor:
+CVI objects are instantiated with empty constructors:
 
 ```julia
-# Davies-Bouldin (DB)
+# Create a Davies-Bouldin (DB) CVI object
 my_cvi = DB()
 ```
 
[…]
