### 2022-10-11 2aea39c
More README simplification
