### 2022-10-11 fcd601f
Rewrite detailed usage, move to docs, remove from readme
