+++ docs/src/man/guide.md
@@ -19,39 +19,63 @@ Pkg.add("ClusterValidityIndices")
 ## Quickstart
 
 This section provides a quick overview of how to use the project.
-For more detailed code usage, please see [Usage](@ref usage).
+For more detailed code usage, please see the [Usage](@ref usage) section.
 For a variety of detailed examples that you can run yourself, please see the [Examples](@ref examples) page.
 
-First, assume that you have a dataset of features/data and labels prescribed by some clustering algorithm:
+First, import the package with:
 
 ```julia
-data_file = "path/to/data.csv"
-data, labels = get_cvi_data(data_file)
+# Import the package
+using ClusterValidityIndices
 ```
 
-All CVI objects in this package are acronymed versions of their full names, which can be found in the [Index](@ref index-types).
-You can create a new CVI structure with a default constructor:
+CVI objects are instantiated with empty constructors:
 
 ```julia
-# Davies-Bouldin (DB)
+# Create a Davies-Bouldin (DB) CVI object
 my_cvi = DB()
 ```
 
-The output of CVIs are called *criterion values*, and they can be computed incrementally with `get_icvi`
+All CVIs are implemented with acronyms of their literature names.
+A list of all of these are found in the [Implemented CVIs](@ref implemented-cvis) section, and their code details can be found in the [Index](@ref index-types).
+
+Next, get data from a clustering process.
+This is a set of samples of features that are clustered and prescribed cluster labels.
+
+> **Note**
+>
+> The `ClusterValidityIndices.jl` package assumes data to be in the form of Float matrices where columns are samples and rows are features.
+> An individual sample is a single vector of features.
+> Labels are vectors of integers where each number corresponds to its own cluster.
 
 ```julia
-n_samples = length(labels)
+# Random data as an example; 10 samples with feature dimenison 3
+dim = 3
+n_samples = 10
+data = rand(dim, n_samples)
+labels = collect(1:n_samples)
+```
+
+The output of CVIs are called *criterion values*, and they can be computed both incrementally and in batch with `get_cvi`.
+Compute in batch by providing a matrix of samples and a vector of labels:
+
+```julia
+criterion_value = get_cvi(my_cvi, data, labels)
+```
+
+or incrementally with the same function by passing one sample and label at a time:
+
+```julia
+# Create a container for the values and iterate
 criterion_values = zeros(n_samples)
 for i = 1:n_samples
-    criterion_values[i] = get_icvi(data[:, i], labels[i])
+    criterion_values[i] = get_cvi(my_cvi, data[:, i], labels[i])
 end
 ```
 
-or in batch with `get_cvi`
-
-```julia
-criterion_value = get_cvi(data, labels)
-```
+> **Note**
+>
+> Each module has a batch and incremental implementation, but `ClusterValidityIndices.jl` does not yet support switching between batch and incremental modes with the same CVI object.
 
 ## [Implemented CVIs](@id implemented-cvis)
 
@@ -71,16 +95,14 @@ The exported constant `CVI_MODULES` also contains a list of these CVIs for conve
 
 ## [Usage](@id usage)
 
-The usage of these CVIs requires an understanding of:
+The usage of these CVIs covers the following:
 
 - [Data](@ref data) assumptions of the CVIs.
 - [How to instantiate](@ref instantiation) the CVIs.
 - [Incremental vs. batch](@ref inc-batch) evaluation.
-- [Updating](@ref updating) internal CVI parameters.
-- [Computing and extracting](@ref criterion-values) the criterion values.
-- [Porcelain functions](@ref porcelain) that are available to simplify operation.
+- [Advanced usage](@ref advanced-usage) for under-the-hood.
 
[…]
