+++ README.md
@@ -53,15 +53,10 @@ Please read the [documentation](https://ap6yc.github.io/ClusterValidityIndices.j
 - [ClusterValidityIndices](#clustervalidityindices)
   - [Table of Contents](#table-of-contents)
   - [Overview](#overview)
-  - [Installation](#installation)
-  - [Quickstart](#quickstart)
-  - [Implemented CVI/ICVIs](#implemented-cviicvis)
+    - [Installation](#installation)
+    - [Quickstart](#quickstart)
+    - [Implemented CVI/ICVIs](#implemented-cviicvis)
     - [Examples](#examples)
-  - [Detailed Usage](#detailed-usage)
-    - [Data](#data)
-    - [Instantiation](#instantiation)
-    - [Incremental vs. Batch](#incremental-vs-batch)
-    - [Advanced Usage](#advanced-usage)
   - [Structure](#structure)
   - [Contributing](#contributing)
   - [Acknowledgements](#acknowledgements)
@@ -79,7 +74,7 @@ This outline begins with [a list of CVIs](#implemented-cviicvis) that are implem
 [Quickstart](#quickstart) provides an overview of how to use this project, while [Structure](#structure) outlines the project file structure, giving context to the locations of every component of the project.
 [Usage](#usage) outlines the general syntax and workflow of the CVIs/ICVIs.
 
-## Installation
+### Installation
 
 This project is distributed as a Julia package, available on [JuliaHub](https://juliahub.com/).
 Its usage follows the usual Julia package installation procedure, interactively:
@@ -103,10 +98,10 @@ julia> ]
 (@v1.8) pkg> add https://github.com/AP6YC/ClusterValidityIndices.jl
 ```
 
-## Quickstart
+### Quickstart
 
 This section provides a quick overview of how to use the project.
-For more detailed code usage, please see [Usage](#usage).
+For more detailed code usage, please see the [Detailed Usage](@ref usage).
 
 First, import the package with:
 
@@ -125,7 +120,7 @@ my_cvi = DB()
 All CVIs are implemented with acronyms of their literature names.
 A list of all of these are found in the [Implemented CVIs/ICVIs](#implemented-cviicvis) section.
 
-Then, get data from a clustering process.
+Next, get data from a clustering process.
 This is a set of samples of features that are clustered and prescribed cluster labels.
 
 > **Note**
@@ -135,7 +130,7 @@ This is a set of samples of features that are clustered and prescribed cluster l
 > Labels are vectors of integers where each number corresponds to its own cluster.
 
 ```julia
-# Random data as an example with ten samples
+# Random data as an example; 10 samples with feature dimenison 3
 dim = 3
 n_samples = 10
 data = rand(dim, n_samples)
@@ -149,7 +144,7 @@ Compute in batch by providing a matrix of samples and a vector of labels:
 criterion_value = get_cvi(my_cvi, data, labels)
 ```
 
-or incrementally by passing one sample and label at a time:
+or incrementally with the same function by passing one sample and label at a time:
 
 ```julia
 # Create a container for the values and iterate
@@ -163,7 +158,7 @@ end
 >
 > Each module has a batch and incremental implementation, but `ClusterValidityIndices.jl` does not yet support switching between batch and incremental modes with the same CVI object.
 
-## Implemented CVI/ICVIs
+### Implemented CVI/ICVIs
 
 This project has implementations of the following CVIs in both batch and incremental variants:
 
@@ -185,113 +180,6 @@ A [basic example](https://ap6yc.github.io/ClusterValidityIndices.jl/dev/getting-
 
 Futhermore, there are a variety of examples in the [Examples](https://ap6yc.github.io/ClusterValidityIndices.jl/dev/examples/) section of the documentation for a variety of use cases of the project.
[…]
