### 2022-10-12 dae4fef
Rework guide installation and TOC
