### 2022-10-12 fd8d4c6
Write about dependencies in contributing guide
