### 2022-10-12 02b35c3
Update basic example to reflect new usage
