### 2022-10-12 8071404
Remove structure from README, prep docs release
