### 2022-10-12 #3527 @AP6YC
@rMassimiliano Release [0.5.0](https://github.com/AP6YC/ClusterValidityIndices.jl/commit/da69f07ecd35d4b258538d8536478339307bbf8a) significantly updates the package and hopefully addresses your concerns. In summary, the following changes have been made:

1. The package usage has simplified to target `get_cvi!` as the main function of the package, categorizing other workflows to "advanced usage". Furthermore, the CVI/ICVI discrepancy is done away with to reduce confusion in favor of referring to every module as a "CVI that can run incrementally or in batch."
2. The README has been simplified to include a functional example, shortened quickstart, removed advanced usage, and fixed links. Furthermore, each implemented CVI now links to the correct listing in the full index in the documentation.
3. The [Documentation](https://ap6yc.github.io/ClusterValidityIndices.jl/dev/) has been similarly overhauled to reflect the [simplified usage](https://ap6yc.github.io/ClusterValidityIndices.jl/dev/man/guide/), which includes the executable DemoCards [examples](https://ap6yc.github.io/ClusterValidityIndices.jl/dev/examples/).
4. A developer index has been added to get every internal docstring into the documentation as well.
5. All types, type fields, and functions have expanded docstrings, which includes the relevant citations for every CVI being in both the [type](https://ap6yc.github.io/ClusterValidityIndices.jl/dev/man/full-index/#ClusterValidityIndices.DB) and [constructor](https://ap6yc.github.io/ClusterValidityIndices.jl/dev/man/full-index/#ClusterValidityIndices.DB-Tuple{}) docstrings.
6. As a cherry on top, nearly all types and functions were reworked in the package to roughly double the package speed (i.e., halving testing time).
