+++ v0.5.0/examples/basic-examples/icvi-example.jl
@@ -1,0 +1,45 @@
+using ClusterValidityIndices    # CVI/ICVI
+using AdaptiveResonance         # DDVFA
+using MLDatasets                # Iris dataset
+using DataFrames                # DataFrames, necessary for MLDatasets.Iris()
+using MLDataUtils               # Shuffling and splitting
+using Printf                    # Formatted number printing
+
+iris = Iris(as_df=false)
+features, labels = iris.features, iris.targets
+
+labels = convertlabel(LabelEnc.Indices{Int}, vec(labels))
+unique(labels)
+
+# Create a Distributed Dual-Vigilance Fuzzy ART (DDVFA) module with default options
+art = DDVFA()
+typeof(art)
+
+# Setup the data configuration for the module
+data_setup!(art.config, features)
+# Verify that the data is setup
+art.config.setup
+
+# Create an ICVI object
+icvi = CH()
+
+# Setup the online/streaming clustering
+n_samples = length(labels)          # Number of samples
+c_labels = zeros(Int, n_samples)    # Clustering labels
+criterion_values = zeros(n_samples) # ICVI outputs
+
+# Iterate over all samples
+for ix = 1:n_samples
+    # Extract one sample
+    sample = features[:, ix]
+    # Cluster the sample online
+    c_labels[ix] = train!(art, sample)
+    # Get the new criterion value (ICVI output)
+    criterion_values[ix] = get_cvi!(icvi, sample, c_labels[ix])
+end
+
+# See the list of criterion values
+criterion_values
+
+# This file was generated using Literate.jl, https://github.com/fredrikekre/Literate.jl
+
