### 2022-10-13 e2e891d
Change guide notes to note blocks
