### 2022-10-13 8441987
Fix basic examples in guide and readme
