### 2022-10-13 23b8d76
Add CVI list section to build
