### 2022-10-13 79df26f
Add blurb with links in cvi list
