### 2022-10-13 8fdad1f
Remove deprecated utils
