+++ src/ClusterValidityIndices.jl
@@ -70,9 +70,6 @@ using Statistics: mean
 # Common structures
 include("common.jl")
 
-# CVI utility functions
-include("utils.jl")
-
 # All of the CVI modules
 include("CVI/CVI.jl")
 
@@ -105,8 +102,6 @@ export
     get_cvi!,
 
     # CVI utilities
-    sort_cvi_data,
-    relabel_cvi_data,
     CVI_MODULES
 
 end
