### 2022-10-13 2dd58d0
Add constructor example to all modules
