### 2022-10-13 2f4a594
Move advanced usage functions to internals, up docs
