### 2022-10-13 e4b4aac
Add plot to icvi example, up dates
