### 2022-10-13 ca0c40a
Update demo descriptions
