+++ docs/examples/cvis/icvi-comparison.jl
@@ -5,7 +5,7 @@
 # date: 2022-10-12
 # author: "[Sasha Petrenko](https://github.com/AP6YC)"
 # julia: 1.8
-# description: This demo illustrates the differing behavior of each ICVI.
+# description: A comparison of the differing behavior of each ICVI.
 # ---
 
 # ## Overview
@@ -50,7 +50,7 @@ typeof(art)
 # Because we are streaming clustering, we must setup the internal data setup of the DDVFA module.
 # This is akin to doing some data preprocessing and communicating the dimension of the data, bounds, etc. to the module beforehand.
 ## Setup the data configuration for the module
-data_setup!(art.config, features)
+data_setup!(art, features)
 ## Verify that the data is setup
 art.config.setup
 
@@ -104,7 +104,13 @@ function plot_cvis(range)
     ## Iterate over the range of ICVI indices provided
     for jx = range
         ## Plot the ICVI criterion values versus sample index
-        plot!(p, 1:n_samples, criterion_values[jx, :], label = string(typeof(icvis[jx])))
+        plot!(
+            p,                              # Modify the plot object
+            1:n_samples,                    # x-axis iteration
+            criterion_values[jx, :],        # y-axis criterion value
+            linewidth=3,                    # Thicken the lines for visibility
+            label=string(typeof(icvis[jx])) # Label is the type of CVI
+        )
     end
     ## Return the plotting object for IJulia display
     return p
