+++ docs/examples/cvis/clustering-comparison.jl
@@ -1,21 +1,21 @@
 # ---
-# title: ICVI Simple Example
-# id: icvi_example
-# cover: assets/icvi-example.png
+# title: Clustering Comparison
+# id: clustering_comparison
+# cover: assets/clustering-comparison.png
 # date: 2022-10-13
 # author: "[Sasha Petrenko](https://github.com/AP6YC)"
 # julia: 1.8
-# description: How to use a CVI incrementally (i.e., ICVI).
+# description: A comparison of two clustering processes with one ICVI.
 # ---
 
 # ## Overview
 
-# This demo is a simple example of how to use CVIs incrementally (ICVI).
-# Here, we load a simple dataset and run a basic clustering algorithm to prescribe a set of clusters to the features.
-# We will take advantage of the fact that we can compute a criterion value at every step by running the ICVI alongside an online clustering algorithm.
-# This simple example demonstrates the usage of a single ICVI, but it may be substituted for any other ICVI in the `ClusterValidityIndices.jl` package.
+# This demo shows how to use CVIs to measure how two online clustering processes can differ.
+# Here, we load a simple dataset and run two clustering algorithms to prescribe a set of clusters to the features.
+# We will also compute their respective criterion values in-the-loop.
+# Though this simple example demonstrates the usage of a single CVI, it may be substituted for any other CVI in the `ClusterValidityIndices.jl` package.
 
-# ## Online Clustering
+# ## Clustering
 
 # ### Data Setup
 
@@ -45,36 +45,44 @@ unique(labels)
 # Because these clustering algorithms run online, we can both cluster and compute a new criterion value at every step.
 # For more on these ART algorithms, see [`AdaptiveResonance.jl`](https://github.com/AP6YC/AdaptiveResonance.jl).
 
-## Create a Distributed Dual-Vigilance Fuzzy ART (DDVFA) module with default options
-art = DDVFA()
-typeof(art)
+# We will create a DDVFA module
+## Create a Distributed Dual-Vigilance Fuzzy ART (DDVFA) module with default options and a DVFA module for comparison
+arts = [
+    DDVFA(),
+    DVFA(),
+]
+typeof(arts)
 
-# Because we are streaming clustering, we must setup the internal data setup of the DDVFA module.
+# Because we are streaming clustering, we must setup the internal data setup of both of the modules.
 # This is akin to doing some data preprocessing and communicating the dimension of the data, bounds, etc. to the module beforehand.
-## Setup the data configuration for the module
-data_setup!(art, features)
-## Verify that the data is setup
-art.config.setup
+## Setup the data configuration for both modules
+for art in arts
+    data_setup!(art, features)
+end
 
 # We can now cluster and get the criterion values online.
-# We will do this by creating an ICVI object, setting up containers for the iterations, and then iterating.
+# We will do this by creating two CVI objects for both clustering modules, setting up containers for the iterations, and then iterating.
 
-## Create an ICVI object
-icvi = CH()
+## Create two CVI objects, one for each clustering module
+n_cvis = length(arts)
+cvis = [CH() for _ = 1:n_cvis]
 
 ## Setup the online/streaming clustering
-n_samples = length(labels)          # Number of samples
-c_labels = zeros(Int, n_samples)    # Clustering labels
-criterion_values = zeros(n_samples) # ICVI outputs
+n_samples = length(labels)              # Number of samples
+c_labels = zeros(Int, n_samples, n_cvis)     # Clustering labels for both
+criterion_values = zeros(n_samples, n_cvis)  # ICVI outputs
 
 ## Iterate over all samples
 for ix = 1:n_samples
     ## Extract one sample
[…]
