+++ docs/examples/basic-examples/cvi-example.jl
@@ -22,7 +22,7 @@
 # First, we must load all of our dependencies.
 # We will load the `ClusterValidityIndices.jl` along with some data utilities and the Julia `Clustering.jl` package to cluster that data.
 using ClusterValidityIndices    # CVI/ICVI
-using Clustering                # DBSCAN
+using Clustering                # Fuzzy C-Means
 using MLDatasets                # Iris dataset
 using DataFrames                # DataFrames, necessary for MLDatasets.Iris()
 using MLDataUtils               # Shuffling and splitting
