### 2022-10-14 b414581
Update new example algorithms
