### 2022-10-14 8548f99
Update ending statements in examples
