+++ docs/examples/basic-examples/icvi-example.jl
@@ -92,7 +92,7 @@ p = plot(
     ylabel = "$(typeof(icvi)) Value",
 )
 
-# Because of the visualization afforded by computing the criterion value incrementally, this plot can tell us several things
+# Because of the visualization afforded by computing the criterion value incrementally, this plot can tell us several things.
 # First, we see that the CVI has a value of zero until the second cluster is encountered, which makes sense because there cannot be measurements of inter-/intra-cluster separation until there is more than one cluster.
 # Second, we see that the criterion value evolves at each time step as the clustering process occurs.
 png("assets/icvi-example") #hide
