### 2022-10-18 7b39bb1
Add version constant export and test
