### 2022-10-18 cb759f0
Refactor constants test
