### 2022-10-25 #3527 @rMassimiliano
@AP6YC good job! I think the updated documentation and examples will be beneficial for new users.

@adi3 my checklist is now complete. I find Release 0.6.0 satisfactory.
