### 2022-11-07 #3527 @adi3
@AP6YC also please look into and fix that invalid DOI shown above in your references
