### 2022-11-10 d59f042
Trim end statement comments
