### 2022-11-10 f95425d
Fix CH bug for zero when n_clusters==2
