### 2022-11-10 49bd01d
Move test data to test folder, dev sklearn test
