### 2022-11-11 8e974e6
Add full scikit-learn test in cvi testset
