+++ test/test_cvis.jl
@@ -12,36 +12,6 @@ A single test set for the testing the functionality of all CVIS modules.
 # TESTSETS
 # --------------------------------------------------------------------------- #
 
-@testset "ScikitLearn Equivalence" begin
-    @info "--- Testing ScikitLearn Equivalence ---"
-    using PyCall, Distributions
-    py_sklearn_metrics = pyimport("sklearn.metrics")
-    # mv_normal1 = MvNormal([0.,0.], [1.5 0.5; 0.5 3])
-    # mv_normal2 = MvNormal([10,8], [1.5 1; 1 2])
-    # mv_normal3 = MvNormal([0,-15], [5 2; 2 5])
-    # mv_normal4 = MvNormal([-20,20], [10 5; 5 10])
-    # rg_nclust = 2:50
-
-    # list_cvi = zeros(length(rg_nclust))
-    # list_cvi_py = zeros(length(rg_nclust))
-
-    # Grab all the data paths for testing
-    data_paths = readdir("data", join=true)
-
-    # Initialize the results data containers
-    data, labels, n_samples = Dict(), Dict(), Dict()
-
-    for data_path in data_paths
-        # Load the data, get a subset, and relabel in order
-        local_data, local_labels = get_cvi_data(data_path)
-
-        # Store the sanitized data
-        data[data_path] = local_data
-        labels[data_path] = local_labels
-        n_samples[data_path] = length(local_labels)
-    end
-end
-
 @testset "CVIs" begin
     @info "--- CVI Testing ---"
 
@@ -67,6 +37,19 @@ end
         data[data_path] = local_data
         labels[data_path] = local_labels
         n_samples[data_path] = length(local_labels)
+    end
+
+    @info "--- Testing ScikitLearn Equivalence ---"
+    py_sklearn_metrics = pyimport("sklearn.metrics")
+    for data_path in data_paths
+        cvi = CH()
+        criterion_value_j = get_cvi!(cvi, data[data_path], labels[data_path])
+        criterion_value_p = (
+            py_sklearn_metrics.calinski_harabasz_score(
+                data[data_path]', labels[data_path]
+            )
+        )
+        @test isapprox(criterion_value_j, criterion_value_p)
     end
 
     # Incremental
