### 2022-11-11 120878c
Add full scikit-learn comparison example
