+++ docs/Project.toml
@@ -4,6 +4,7 @@ ClusterValidityIndices = "2fefd308-f647-4698-a2f0-e90cfcbca9ad"
 Clustering = "aaaa29a8-35af-508c-8bc3-b662a17a0fe5"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 DemoCards = "311a05b2-6137-4a5a-b473-18580a3d38b5"
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 LiveServer = "16fef848-5104-11e9-1b77-fb7a48bbb589"
