### 2022-11-11 013b8a6
Debug new example, save image
