+++ docs/examples/cvis/scikit-learn.jl
@@ -1,7 +1,7 @@
 # ---
 # title: ScikitLearn Comparison
 # id: scikit-learn
-# cover: assets/icvi-comparision.png
+# cover: assets/scikit-learn.png
 # date: 2022-11-11
 # author: "[Sasha Petrenko](https://github.com/AP6YC)"
 # julia: 1.8
@@ -62,7 +62,6 @@ end
 X = reduce(hcat, [rand(dists[i], n_samples_per) for i=1:n_distributions])
 p = plot()
 scatter!(p, X[1, :], X[2, :])
-display(p)
 
 # ## Cluster and Compute Metrics
 
@@ -74,17 +73,17 @@ display(p)
 # Here, we will cluster the k-means while testing a range of seed values, comparing the equivalence of the Calinski-Harabasz (CH) metric between this package and the scikit-learn implementation.
 ## Create a range of k-mean seeds to test across along with CVI containers
 km_range = 2:n_k_means
-criterion_values_jl = zeros(length(rg_nclust))
-criterion_values_py = zeros(length(rg_nclust))
+criterion_values_jl = zeros(length(km_range))
+criterion_values_py = zeros(length(km_range))
 for (i, n_clust) = enumerate(km_range)
     ## Get the labels prescribed k-means to each sample
     labels_km = kmeans(X, n_clust).assignments
 
-    # Create a Calinski-Harabasz CVI and get its criterion value
+    ## Create a Calinski-Harabasz CVI and get its criterion value
     cvi = CH()
     criterion_values_jl[i] = get_cvi!(cvi, X, labels_km)
 
-    # Compute the criterion value with the scikit-learn implementation
+    ## Compute the criterion value with the scikit-learn implementation
     criterion_values_py[i] = skmetrics.calinski_harabasz_score(X', labels_km)
 end
 ## Show that the values are all equivalent between implementations
@@ -92,9 +91,10 @@ isapprox(criterion_values_jl, criterion_values_py)
 
 # We can visualize the result as well by plotting the trendlines of the criterion values from each implementation.
 q = plot(legend=:bottomright)
-plot!(q, rg_nclust, list_cvi, marker=:d)
+plot!(q, km_range, criterion_values_jl, marker=:d)
+plot!(q, km_range, criterion_values_py, marker=:d)
 ylabel!(q, "Calinski-Harabasz CVI")
 xlabel!(q, "K-Means Seed")
-display(q)
 
-
+# After visualizing the results and seeing that the trendlines are equivalent, we also see that unsurprisingly the best k-means seed value is k=4, since we know that
+png("assets/scikit-learn") #hide
