### 2022-11-14 bfc1c1c
Add scikitlearn install in CI deps
