+++ .github/workflows/CI.yml
@@ -44,8 +44,12 @@ jobs:
         uses: styfle/cancel-workflow-action@0.6.0
         with:
           access_token: ${{ github.token }}
-
       - uses: actions/checkout@v2
+      # Setup python dependencies
+      - name: Install scikit-learn
+        run: |
+          python -m pip install --upgrade pip
+          pip install -U scikit-learn
       - uses: julia-actions/setup-julia@v1
         with:
           version: ${{ matrix.version }}
