### 2022-11-15 c446cd3
Fix sklearn install on windows and mac for pycall
