+++ .github/workflows/CI.yml
@@ -38,6 +38,7 @@ jobs:
       # Set this to download test data outside of repo
       DATADEPS_ALWAYS_ACCEPT: true
     steps:
+    #  styfle/cancel-workflow-action@0.6.0, actions/checkout@v2, actions/cache@v2, codecov/codecov-action@v1, coverallsapp/github-action@master
       # Cancel ongoing tests if pushing to branch again before the previous
       # build is finished.
       - name: Cancel ongoing tests for previous commits
