### 2022-11-15 6f1ef3f
Add catch for sklearn versions missing CH metric
