### 2022-11-16 #3527 @AP6YC
@adi3 I fixed the DOI issue and have created an archived release; checking the DOI now:
