### 2022-11-16 #3527 @AP6YC
@adi3 Below is the information that you requested:

- DOI: 10.5281/zenodo.7327501
- Archive link: https://zenodo.org/record/7327501
- Version: v0.6.2
