### 2022-11-21 #3527 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
